% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMethods.R
\name{annotate,EdgeObject,character,missing-method}
\alias{annotate,EdgeObject,character,missing-method}
\title{Annotate an EdgeObject, without checking the target}
\usage{
\S4method{annotate}{EdgeObject,character,missing}(object, target)
}
\arguments{
\item{object}{An EdgeObject}

\item{target}{Character, target of annotation}
}
\value{
An annotated \code{EdgeObject}.
}
\description{
Annotate an EdgeObject, without checking the target
}
