% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{annotation,EdgeObject-method}
\alias{annotation,EdgeObject-method}
\title{Get annotation information from an EdgeObject}
\usage{
\S4method{annotation}{EdgeObject}(object)
}
\arguments{
\item{object}{An \code{EdgeObject}.}
}
\value{
A character string, indicating the annotation type
}
\description{
Get annotation information from an EdgeObject
}
