% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoT.R
\name{appendPseudoT}
\alias{appendPseudoT}
\title{Append degree of freedom and pseudo t-statistics to dgeTable}
\usage{
appendPseudoT(edgeResult, dgeTable)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{dgeTable}{A \code{data.frame}, derived from \code{dgeTables} or \code{dgeTable} usually.}
}
\value{
A new \code{data.frame} with two new columns, \code{df} and \code{pseudoT}, containing degree of freem and the (pseudo) t-statistic, respective.y
The function relies on the fact that the degree of freedom (`df.residual` in GLM result) is the same for all genes.
If this is not the case, it will use the `FeatureName` column in gene annotation to match the degree of freedoms.
The function is only used internally by \code{dgeTablesWithPseudoT} and \code{dgeTableWithPseudoT}.
}
\description{
Append degree of freedom and pseudo t-statistics to dgeTable
}
\note{
the \code{lower.tail} option in \code{qt} is not vectorized, therefore the sign should be provided separately.
}
