% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockdown.R
\name{appendRanks}
\alias{appendRanks}
\title{Append ranks to dgeTbl}
\usage{
appendRanks(dgeTbl)
}
\arguments{
\item{dgeTbl}{A dgeTbl, a data.frame that at least contain following columns
\code{logFC} and \code{PValue}}
}
\value{
The dgeTble with four columns appended: \code{rank_logFC},
\code{rank_PValue}, \code{rank_absLogFC}, and \code{total_features}, and
sorted by increasing P-values
}
\description{
Append ranks to dgeTbl
}
\examples{
myTbl <- data.frame(GeneSymbol=LETTERS[1:5], 
  logFC=rnorm(5), PValue=runif(5, 0, 1))
appendRanks(myTbl)
}
