% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendZScore.R
\name{appendZScore}
\alias{appendZScore}
\title{Append zscores to dgeTable}
\usage{
appendZScore(dgeTable)
}
\arguments{
\item{dgeTable}{A \code{data.frame}, derived from \code{dgeTables} or \code{dgeTable} usually. It must contain the following two columns: \code{PValue}, and \code{logFC} (case sensitive).}
}
\value{
A new \code{data.frame} with one new column, \code{zScore}, containing the z-score transformed from p-values. If that column exists, it will be rewritten.
The function is similar to \code{appendPseudoT}, with the difference that the Gaussian distribution is used instead of the t-distribution. This solution
delivers less extreme values, because t-distribution is heavy-tailed. It allows comparison between studies of different sample sizes.
}
\description{
Append zscores to dgeTable
}
\seealso{
\code{\link{appendPseudoT}}
}
