% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{assertEdgeToptable}
\alias{assertEdgeToptable}
\title{Assert that the input data.frame is a valid EdgeTopTable}
\usage{
assertEdgeToptable(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
Logical
}
\description{
Assert that the input data.frame is a valid EdgeTopTable
}
