% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{aveExpr}
\alias{aveExpr}
\title{Get aveExpr threshold in LimmaSigFilter}
\usage{
aveExpr(limmaSigFilter)
}
\arguments{
\item{limmaSigFilter}{An \code{EdgeSigFilter} object}
}
\value{
Numeric values of the logCPM filter
}
\description{
Get aveExpr threshold in LimmaSigFilter
}
