% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{BCV}
\alias{BCV}
\alias{BCV,DGEList-method}
\alias{BCV,EdgeResult-method}
\title{Return a data.frame of BCV values}
\usage{
BCV(x)

\S4method{BCV}{DGEList}(x)

\S4method{BCV}{EdgeResult}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A \code{data.frame} of BCV values.
}
\description{
Return a data.frame of BCV values
}
\section{Methods (by class)}{
\itemize{
\item \code{BCV(DGEList)}: Method for DGEList

\item \code{BCV(EdgeResult)}: Method for EdgeResult

}}
