\name{blackyellow}
\alias{blackyellow}
\alias{yellowblack}
\alias{whiteblue}
\alias{whitered}
\alias{blackred}
\alias{blackgreen}
\title{Two-color panels}
\description{
  Precompiled two-color panels for visualization purposes
}
\usage{
blackyellow(n)
yellowblack(n)
whiteblue(n)
whitered(n)
blackred(n)
blackgreen(n)
}
\arguments{
  \item{n}{Number of colors needed}
}
\details{
  See function definitions
}
\value{Character vector of length \code{n} coding colors}
\author{Jitao David Zhang}
\seealso{
  \code{royalbluered} for three-color systems
}
\examples{
display.twocolor.panels()
}

