% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{boxplot,EdgeObject-method}
\alias{boxplot,EdgeObject-method}
\title{Boxplot of an EdgeObject}
\usage{
\S4method{boxplot}{EdgeObject}(
  x,
  type = c("normFactors", "modLogCPM"),
  xlab = "",
  ylab = NULL,
  main = "",
  ...
)
}
\arguments{
\item{x}{An EdgeObject}

\item{type}{The type of boxplot: 'normFactors' and 'modLogCPM' are supported.}

\item{xlab}{Character, xlab.}

\item{ylab}{Character, ylab.}

\item{main}{Character, title.}

\item{...}{Passed to \code{boxplot}.}
}
\value{
Called for its side effect of plotting; returns invisibly NULL.
}
\description{
Boxplot of an EdgeObject
}
