% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcNormFactorsIfNot}
\alias{calcNormFactorsIfNot}
\title{Calculate normalisation factor if not}
\usage{
calcNormFactorsIfNot(dgeList)
}
\arguments{
\item{dgeList}{A \code{DGEList} object
Calculate the normalisation factors if not done yet}
}
\value{
Updated \code{dgeList} object with \code{norm.factors} filled
}
\description{
Calculate normalisation factor if not
}
