% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{cameraScore}
\alias{cameraScore}
\title{Return cameraScore}
\usage{
cameraScore(pvalue, direction)
}
\arguments{
\item{pvalue}{PValues of camera result}

\item{direction}{directions of camera result, either 'Up' (positive) or 'Down' (negative)}
}
\value{
A numeric vector, containing camera Scores
}
\description{
Return cameraScore
}
\examples{
ribiosNGS:::cameraScore(pvalue=c(1, 0.1, 0.01), direction=c("Up", "Down", "Up"))
}

