% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-slurm-lsf.R
\name{checkContrastNames}
\alias{checkContrastNames}
\title{Check a contrast matrix to make sure that it is likely o.k.}
\usage{
checkContrastNames(contrastMatrix, action = c("message", "warning", "error"))
}
\arguments{
\item{contrastMatrix}{A contrast matrix}

\item{action}{Character strings, the action to perform in case the names show irregularities

Right now, the function checks no column names contain the equal sign.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of issuing
  a message, warning, or error if column names contain equal signs.
}
\description{
Check a contrast matrix to make sure that it is likely o.k.
}
\examples{
testDesign <- cbind(Control=rep(1,8), Treatment=rep(c(0,1),4), Batch=rep(c(0, 1), each=4))
problemContrast <- limma::makeContrasts("Treatment"="Treatment",
  "Batch=Batch", ## problematic
  levels=testDesign)
checkContrastNames(problemContrast, action="message")
if(requireNamespace("testthat")) {
  testthat::expect_warning(checkContrastNames(problemContrast,
         action="warning"))
  testthat::expect_error(checkContrastNames(problemContrast,
         action="error"))
}
}
