% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{commonBCV}
\alias{commonBCV}
\alias{commonBCV,DGEList-method}
\alias{commonBCV,EdgeResult-method}
\title{Common biological coefficients of variance (BCV)}
\usage{
commonBCV(x)

\S4method{commonBCV}{DGEList}(x)

\S4method{commonBCV}{EdgeResult}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A numeric value of the common BCV.
}
\description{
Common biological coefficients of variance (BCV)
}
\section{Methods (by class)}{
\itemize{
\item \code{commonBCV(DGEList)}: method for DGEList

\item \code{commonBCV(EdgeResult)}: method for EdgeResult

}}
