% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{commonDisp}
\alias{commonDisp}
\alias{commonDisp,DGEList-method}
\alias{commonDisp,EdgeObject-method}
\title{Common dispersion}
\usage{
commonDisp(object)

\S4method{commonDisp}{DGEList}(object)

\S4method{commonDisp}{EdgeObject}(object)
}
\arguments{
\item{object}{An object}
}
\value{
A numeric value of the common dispersion.
}
\description{
Common dispersion
}
\section{Methods (by class)}{
\itemize{
\item \code{commonDisp(DGEList)}: Method for DGEList

\item \code{commonDisp(EdgeObject)}: Method for EdgeObject

}}
