% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastMatrix,EdgeObject-method}
\alias{contrastMatrix,EdgeObject-method}
\title{Extract contrast matrix from an EdgeObject object}
\usage{
\S4method{contrastMatrix}{EdgeObject}(object)
}
\arguments{
\item{object}{An EdgeObject object}
}
\value{
A contrast matrix.
}
\description{
Extract contrast matrix from an EdgeObject object
}
