% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastMatrix,EdgeResult-method}
\alias{contrastMatrix,EdgeResult-method}
\title{Extract contrast matrix from an EdgeResult object}
\usage{
\S4method{contrastMatrix}{EdgeResult}(object)
}
\arguments{
\item{object}{An EdgeResult object}
}
\value{
A contrast matrix.
}
\description{
Extract contrast matrix from an EdgeResult object
}
