% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{contrastMatrix-set}
\alias{contrastMatrix-set}
\alias{contrastMatrix<-}
\alias{contrastMatrix<-,EdgeObject,matrix-method}
\title{Assign contrast matrix}
\usage{
contrastMatrix(object) <- value

\S4method{contrastMatrix}{EdgeObject,matrix}(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Matrix}
}
\value{
The updated object with the new contrast matrix.
}
\description{
Assign contrast matrix
}
\section{Functions}{
\itemize{
\item \code{contrastMatrix(object = EdgeObject) <- value}: Method for EdgeObject

}}
