% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastSampleIndices,EdgeResult,integer-method}
\alias{contrastSampleIndices,EdgeResult,integer-method}
\title{Extract contrast sample indices}
\usage{
\S4method{contrastSampleIndices}{EdgeResult,integer}(object, contrast)
}
\arguments{
\item{object}{An EdgeResult object.}

\item{contrast}{Character, indicating the contrast of interest.}
}
\value{
A list of integer vectors indicating sample indices for each group in the contrast.
}
\description{
Extract contrast sample indices
}
