% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq.R
\name{convertPDF2PNG}
\alias{convertPDF2PNG}
\title{Convert PDF files to PNG files}
\usage{
convertPDF2PNG(object)
}
\arguments{
\item{object}{A riboSeqAnalysis object}
}
\value{
A character string vector of output files
}
\description{
Convert PDF files to PNG files
}
