% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{countByGroup}
\alias{countByGroup}
\alias{maxCountByGroup}
\alias{hasNoReplicate}
\title{Sample counts by group}
\usage{
countByGroup(edgeObj)

maxCountByGroup(edgeObj)

hasNoReplicate(edgeObj)
}
\arguments{
\item{edgeObj}{An EdgeObject object}
}
\value{
A named vector if integers, sample counts by group
}
\description{
Sample counts by group
}
\section{Functions}{
\itemize{
\item \code{maxCountByGroup}: Returns the max count

\item \code{hasNoReplicate}: Returns \code{TRUE} if the largest group has only one sample
}}

