% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CountDgeResult-class}
\alias{CountDgeResult-class}
\title{Object that contains count data, dgeTables, and sigFilter}
\description{
Object that contains count data, dgeTables, and sigFilter
}
\section{Slots}{

\describe{
\item{\code{dgeTables}}{A list of dgeTable}

\item{\code{sigFilter}}{Significantly regulated gene filter}
}}

\note{
The object is used only for inheritance
}
