% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{counts,DGEList-method}
\alias{counts,DGEList-method}
\title{Return counts in a DGEList object}
\usage{
\S4method{counts}{DGEList}(object)
}
\arguments{
\item{object}{A \code{DGEList} object.}
}
\value{
A numeric matrix of counts.
}
\description{
Return counts in a DGEList object
}
