% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{countsSVA}
\alias{countsSVA}
\title{Apply SVA to transformed count data}
\usage{
countsSVA(counts, designMatrix, transformFunc = function(counts, designMatrix)
  voom(counts, designMatrix)$E, ...)
}
\arguments{
\item{counts}{A matrix of counts}

\item{designMatrix}{Design matrix}

\item{transformFunc}{A function to transform the count data}
}
\value{
The SV matrix
}
\description{
Apply SVA to transformed count data
}
\examples{
set.seed(1887)
exCounts <- matrix(rpois(12000, 10), nrow=2000, ncol=6)
exCounts[1:100, 2:3] <- exCounts[1:100,2:3]+20
exDesign <- model.matrix(~gl(2,3))
countsSVA(exCounts, designMatrix=exDesign)
}
