% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cpm.EdgeObject}
\alias{cpm.EdgeObject}
\title{cpm for EdgeObject}
\usage{
\method{cpm}{EdgeObject}(y, ...)
}
\arguments{
\item{y}{An EdgeObject object}

\item{...}{Passed to \code{cpm}}
}
\value{
A numeric matrix of counts per million values.
}
\description{
cpm for EdgeObject
}
\seealso{
\code{\link[edgeR]{cpm}}
}
