% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{cpmFilter}
\alias{cpmFilter}
\title{Filter by counts per million (cpm)}
\usage{
cpmFilter(object)
}
\arguments{
\item{object}{An object}
}
\value{
The filtered object.
}
\description{
Filter by counts per million (cpm)
}
