% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{cpmRNA}
\alias{cpmRNA}
\alias{cpmRNA,EdgeObject-method}
\title{Get RNA counts per million (cpm)}
\usage{
cpmRNA(object, ...)

\S4method{cpmRNA}{EdgeObject}(object, ...)
}
\description{
Get RNA counts per million (cpm)
}
\section{Methods (by class)}{
\itemize{
\item \code{EdgeObject}: Method for EdgeObject
}}

