% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{cpmSVA}
\alias{cpmSVA}
\title{Apply SVA to cpm-transformed count data}
\usage{
cpmSVA(counts, designMatrix)
}
\arguments{
\item{counts}{A matrix of counts}

\item{designMatrix}{Design matrix}
}
\value{
The SV matrix
}
\description{
Apply SVA to cpm-transformed count data
}
\examples{
exCounts <- matrix(rpois(12000, 10), nrow=2000, ncol=6)
exCounts[1:100, 2:3] <- exCounts[1:100,2:3]+20
exDesign <- model.matrix(~gl(2,3))
cpmSVA(exCounts, designMatrix=exDesign)
}
