% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{cpmVsn2}
\alias{cpmVsn2}
\title{Fit the vsn model to log2CPM data}
\usage{
cpmVsn2(x, prior.count = 2, normalized.lib.sizes = TRUE,
  verbose = FALSE, ...)
}
\arguments{
\item{x}{A matrix or DGEList object}

\item{prior.count}{Integer, passed to \code{\link[edgeR]{cpm}}}

\item{normalized.lib.sizes}{Logical, passed to \code{\link[edgeR]{cpm}}}

\item{verbose}{Logical, whether diagnostic information is printed for vsn}

\item{...}{Other parameters passed to \code{\link[vsn]{justvsn}}}
}
\value{
A matrix of vsn-transformed expression data
}
\description{
Fit the vsn model to log2CPM data
}
\examples{

myCounts <- matrix(rnbinom(1000, 10, 0.5), nrow=100)
myDgeList <- DGEList(counts=myCounts,
  samples=data.frame(group=gl(5,2)))
  
myCpmVsn2MatRes <- cpmVsn2(myCounts)
myCpmVsn2DGEListRes <- cpmVsn2(myDgeList)

}
