% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{customSmearPlot}
\alias{customSmearPlot}
\title{Custom smear plot}
\usage{
customSmearPlot(
  tbl,
  main,
  xlab,
  ylab,
  pch = 19,
  cex = 0.2,
  smearWidth = 0.5,
  panel.first = grid(),
  smooth.scatter = FALSE,
  lowess = FALSE,
  ...
)
}
\arguments{
\item{tbl}{A \code{data.frame}}

\item{main}{Character string, title of the plot}

\item{xlab}{Character string, xlab}

\item{ylab}{Character string, ylab}

\item{pch}{Point symbol}

\item{cex}{Font size}

\item{smearWidth}{Smear with}

\item{panel.first}{Passed to \code{\link[edgeR]{maPlot}}.}

\item{smooth.scatter}{Passed to \code{\link[edgeR]{maPlot}}.}

\item{lowess}{Passed to \code{\link[edgeR]{maPlot}}.}

\item{...}{Passed to \code{\link[edgeR]{maPlot}}.}
}
\value{
Called for its side effect of plotting; returns invisibly NULL.
}
\description{
Custom smear plot
}
