% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{designMatrix,EdgeObject-method}
\alias{designMatrix,EdgeObject-method}
\title{Extract design matrix from an EdgeObject object}
\usage{
\S4method{designMatrix}{EdgeObject}(object)
}
\arguments{
\item{object}{An EdgeObject object}
}
\value{
A design matrix.
}
\description{
Extract design matrix from an EdgeObject object
}
