% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{designMatrix-set}
\alias{designMatrix-set}
\alias{designMatrix<-}
\alias{designMatrix<-,EdgeObject,matrix-method}
\title{Assign the design matrix}
\usage{
designMatrix(object) <- value

\S4method{designMatrix}{EdgeObject,matrix}(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Matrix}
}
\value{
The updated object with the new design matrix.
}
\description{
Assign the design matrix
}
\section{Functions}{
\itemize{
\item \code{designMatrix(object = EdgeObject) <- value}: Method for EdgeObject

}}
