% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{dgeGML}
\alias{dgeGML}
\title{Return the dgeGML method}
\usage{
dgeGML(edgeResult)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object.}
}
\value{
A \code{DGEGLM} object.
}
\description{
Return the dgeGML method
}
