% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{dgeList}
\alias{dgeList}
\alias{dgeList,EdgeObject-method}
\alias{dgeList,EdgeResult-method}
\title{Extract DGEList from an EdgeObject}
\usage{
dgeList(object)

\S4method{dgeList}{EdgeObject}(object)

\S4method{dgeList}{EdgeResult}(object)
}
\arguments{
\item{object}{An EdgeResult object}
}
\value{
A \code{DGEList} object.
}
\description{
Extract DGEList from an EdgeObject

Extract DGEList from an EdgeResult object
}
\section{Methods (by class)}{
\itemize{
\item \code{dgeList(EdgeObject)}: Extract DGEList from EdgeObject

\item \code{dgeList(EdgeResult)}: Extract DGEList from EdgeResult

}}
