% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{DGEList2}
\alias{DGEList2}
\title{Construct a DGEList2 object}
\usage{
DGEList2(...)
}
\arguments{
\item{...}{A list of DGEList2 objects, can be passed as individual objects
or in a list}
}
\description{
Construct a DGEList2 object
}
