% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{dgeListCamera}
\alias{dgeListCamera}
\title{Perform camera using DGEList}
\usage{
dgeListCamera(dgeList, index, design, contrasts)
}
\arguments{
\item{dgeList}{A DGEList object, with GeneSymbol available}

\item{index}{List of integer indices of genesets, names are namese of gene
sets}

\item{design}{Design matrix}

\item{contrasts}{Contrast matrix}
}
\description{
Perform camera using DGEList
}
