% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DgeResult-class}
\alias{DgeResult-class}
\title{Object that contains input data, dgeTables, and sigFilter}
\description{
Object that contains input data, dgeTables, and sigFilter
}
\section{Slots}{

\describe{
\item{\code{dgeTables}}{A list of dgeTable}

\item{\code{sigFilter}}{Significantly regulated gene filter

The object is used only for inheritance}
}}

