% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeTable.R
\name{dgeTable}
\alias{dgeTable}
\title{Return the top table in a unified format}
\usage{
dgeTable(edgeResult, contrast = NULL)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{contrast}{Character, contrast name of interest. If \code{NULL}, 
all tables are returned in a rbind-form.}
}
\value{
A \code{data.frame}
}
\description{
Return the top table in a unified format
}
