% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeTable.R
\name{dgeTableList}
\alias{dgeTableList}
\title{Return the top tables of specified contrast(s) in a list}
\usage{
dgeTableList(edgeResult, contrast = NULL)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{contrast}{Character, contrast name(s) of interest. If \code{NULL}, 
all tables are returned in a rbind-form.}
}
\value{
A list of \code{data.frame}s
}
\description{
Return the top tables of specified contrast(s) in a list
}
