% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoT.R
\name{dgeTablesWithPseudoT}
\alias{dgeTablesWithPseudoT}
\title{Append dgeTables with pseudo t-statistic}
\usage{
dgeTablesWithPseudoT(edgeResult)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object.}
}
\value{
Similar as \code{dgeTables}, a list of \code{data.frame}, but with additional columns \code{df} (degree of freedom) and \code{pseudoT}.
The pseudo t-statistic is calculated based on the P-value of the likelihood ratio test and the residual degree of freedom by the function. \code{\link{qt}}, and its sign is given by the sign of logFC.
}
\description{
Append dgeTables with pseudo t-statistic
}
\seealso{
\code{\link{dgeTableWithPseudoT}}
}
