% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendZScore.R
\name{dgeTablesWithZScore}
\alias{dgeTablesWithZScore}
\title{Append dgeTables with z-scores}
\usage{
dgeTablesWithZScore(edgeResult)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object.}
}
\value{
Similar as \code{dgeTables}, a list of \code{data.frame}, but with an additional column \code{zScore}.
}
\description{
Append dgeTables with z-scores
}
\seealso{
\code{\link{appendZScore}}, \code{\link{dgeTableWithZScore}}, \code{\link{dgeTableWithPseudoT}}
}
