% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoT.R
\name{dgeTableWithPseudoT}
\alias{dgeTableWithPseudoT}
\title{Append dgeTable with pseudo t-statistic}
\usage{
dgeTableWithPseudoT(edgeResult, contrast = NULL)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object.}

\item{contrast}{A character string, or integer index, or \code{NULL}, to specify the contrast. If \code{NULL}, results of all contrasts are returned.}
}
\value{
Similar as \code{dgeTable}, a \code{data.frame}, but with additional columns \code{df} (degree of freedom) and \code{pseudoT}.
The pseudo t-statistic is calculated based on the P-value of the likelihood ratio test and the residual degree of freedom by the function \code{\link{qt}}, and its sign is given by the sign of logFC.
}
\description{
Append dgeTable with pseudo t-statistic
}
\seealso{
\code{\link{dgeTablesWithPseudoT}}, \code{\link{dgeTable}}
}
