% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dge.R
\name{dgeWithEdgeR}
\alias{dgeWithEdgeR}
\title{Perform differential gene expression analysis with edgeR}
\usage{
dgeWithEdgeR(edgeObj)
}
\arguments{
\item{edgeObj}{An object of \code{EdgeObject}

The function performs end-to-end differential gene expression (DGE) analysis
with common best practice using edgeR}
}
\value{
An \code{EdgeResult} object
}
\description{
Perform differential gene expression analysis with edgeR
}
\examples{

exMat <- matrix(rpois(120, 10), nrow=20, ncol=6)
exGroups <- gl(2,3, labels=c("Group1", "Group2"))
exDesign <- model.matrix(~0+exGroups)
exContrast <- matrix(c(-1,1), ncol=1, dimnames=list(c("Group1", "Group2"), c("Group2.vs.Group1")))
exDescon <- DesignContrast(exDesign, exContrast, groups=exGroups)
exFdata <- data.frame(Identifier=sprintf("Gene\%d", 1:nrow(exMat)))
exPdata <- data.frame(Name=sprintf("Sample\%d", 1:ncol(exMat)),
                     Group=exGroups)
exObj <- EdgeObject(exMat, exDescon, 
                     fData=exFdata, pData=exPdata)
exDgeRes <- dgeWithEdgeR(exObj)
dgeTable(exDgeRes)

}
