% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{dim.EdgeResult}
\alias{dim.EdgeResult}
\title{Dimensions of an EdgeResults}
\usage{
\method{dim}{EdgeResult}(x)
}
\arguments{
\item{x}{An EdgeResult object}
}
\value{
An integer vector of length two (features, samples).
}
\description{
Dimensions of an EdgeResults
}
