\name{dummy-package}
\alias{dummy-package}
\alias{dummy}
\docType{package}
\title{
  dummy
}
\description{
  dummy package to study R
}
\details{
\tabular{ll}{
Package: \tab dummy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-16\cr
License: \tab LGPL-2 \cr
LazyLoad: \tab yes\cr
}
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
Maintainer: Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{ package }
%%\examples{
%%
%%}
