% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edgeBuild}
\alias{edgeBuild}
\title{Build DGEList object ready for generalized linear model fitting}
\usage{
edgeBuild(x, design, robust = FALSE)
}
\arguments{
\item{x}{a numerical matrix or any object that can be transformed as such}

\item{design}{design matrix (subject to dimensionality check)}

\item{robust}{Whether robust estimation (Zhou \emph{et al.}, NAR 2014) should be used}
}
\value{
A DGEList object with dispersions estimated, ready for \code{glmFit}.
}
\description{
The function combines the step of object initialization, library size calculation, and dispersion estimation
}
\note{
robust=TRUE is only valid for edgeR version >= 3.6.0
}
\examples{
expression <- matrix(rnbinom(10000,mu=5,size=2),ncol=4)
design <- cbind(baseline=c(1,1,1,1), x=c(1,1,-1,-1))
edgeObj <- edgeBuild(expression, design)
}
\seealso{
\code{glmFit} to fit a generalized linear model
}

