% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/EdgeObject.R
\docType{class}
\name{EdgeObject-class}
\alias{EdgeObject-class}
\alias{EdgeObject,matrix,DesignContrast-method}
\alias{EdgeObject,FeatAnnoExprs,DesignContrast-method}
\title{EdgeObject argumenting DGEList by including designContrast information}
\usage{
\S4method{EdgeObject}{matrix,DesignContrast}(
  object,
  designContrast,
  fData = NULL,
  pData = NULL,
  remove.zeros = FALSE
)

\S4method{EdgeObject}{FeatAnnoExprs,DesignContrast}(object, designContrast, pData = NULL, remove.zeros = FALSE)
}
\description{
EdgeObject argumenting DGEList by including designContrast information
}
\section{Methods (by generic)}{
\itemize{
\item \code{EdgeObject}: The method for matrix as input

\item \code{EdgeObject}: The method for FeatAnnoExprs as input
}}

\section{Slots}{

\describe{
\item{\code{dgeList}}{A \code{DGEList} object}

\item{\code{designContrast}}{A \code{designContrast} object}
}}

