% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeObject.R
\name{EdgeObject.GctMatrix}
\alias{EdgeObject.GctMatrix}
\title{S3 method to create an EdgeObject from GctMatrix object}
\usage{
EdgeObject.GctMatrix(
  object,
  designContrast,
  fData = NULL,
  pData = NULL,
  remove.zeros = FALSE
)
}
\arguments{
\item{object}{A GctMatrix object, probably from \code{\link[ribiosIO]{read_gct_matrix}}.}

\item{designContrast}{A \code{DesignContrast} object.}

\item{fData}{Feature data}

\item{pData}{Pheno data}

\item{remove.zeros}{Logical
See \code{EdgeObject} for help}
}
\description{
S3 method to create an EdgeObject from GctMatrix object
}
