% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EdgeResult-class}
\alias{EdgeResult-class}
\title{Return a list of differential gene expression tables}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}
}
\value{
A list of \code{data.frame}s, each containing the DGEtable for one
contrast.
}
\description{
Return a list of differential gene expression tables
}
\seealso{
\code{dgeTable} which returns one \code{data.frame} for one or more
given contrasts.
}
