% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{EdgeResult}
\alias{EdgeResult}
\title{Return a list of differential gene expression tables}
\usage{
EdgeResult(edgeObj, dgeGLM, dgeTables)
}
\arguments{
\item{edgeObj}{An EdgeObject}

\item{dgeGLM}{A DGEGLM object}

\item{dgeTables}{A list of DGEtables.}
}
\value{
An \code{EdgeResult object}
}
\description{
Return a list of differential gene expression tables
}
