% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edgeRun}
\alias{edgeRun}
\title{Identify differentially expressed genes using generalized linear models implemented by edgeR}
\usage{
edgeRun(x, design, contrasts, robust = FALSE)
}
\arguments{
\item{x}{A matrix of read counts, or any objects that can be converted to a matrix}

\item{design}{Design matrix (subject to dimensionality check)}

\item{contrasts}{Contrast matrix (subject to dimensionality check)}

\item{robust}{Logical, whether robust dispersion estimation should be used or not}
}
\value{
A list of data frames reporting differntial gene expression
}
\description{
The function provides a simple interface to run differential gene analysis using edgeR using default settings
Users desiring higher flexibility may call functions individually
It is user's responsibility to understand the assumptions behind the default settings and to judge whether they are suitable for the data
}
\note{
robust=TRUE is only valid for edgeR version >= 3.6.0
}
\examples{
expression <- matrix(rnbinom(10000,mu=5,size=2),ncol=4)
design <- cbind(baseline=c(1,1,1,1), x=c(1,1,-1,-1), y=c(-1,1,-1,1))
contrast <- cbind(C1=c(0,1,0), C2=c(0,1,-1))
edgeDiff <- edgeRun(expression, design, contrast, robust=FALSE)
}
\seealso{
\code{\link{edgeBuild}} to build a DGEList object and to estimate dispersions
\code{\link{glmFit}} to fit a DGEList object against the design matrix
\code{\link{edgeTest}} to derive lists of differentially expressed genes

#'
}

