% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EdgeSigFilter-class}
\alias{EdgeSigFilter-class}
\title{EdgeSigFilter: EdgeR result filter for significantly regulated genes}
\description{
EdgeSigFilter: EdgeR result filter for significantly regulated genes
}
\section{Slots}{

\describe{
\item{\code{posLogFC}}{Numeric, positive logFC threshold (larger values are kept)}

\item{\code{negLogFC}}{Numeric, negative logFC threshold (more negative values 
are kept)}

\item{\code{aveExpr}}{Numeric, threshold of average expression (larger values are kept)}

\item{\code{LR}}{Numeric, likelihood ratio (LR) threshold (larger values are kept)}

\item{\code{pValue}}{Numeric, p-value treshold (smaller values are kept)}

\item{\code{FDR}}{Numeric, FDR treshold}
}}

