% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edgeTest}
\alias{edgeTest}
\title{Test differentially expressed genes using edgeR}
\usage{
edgeTest(fit, contrasts)
}
\arguments{
\item{fit}{An \code{DGEGLM} object produced by \code{glmFit}}

\item{contrasts}{a contrast matrix. If a vector is given, it is converted into a column vector.}
}
\value{
A list of data frames containing information about differential expression
}
\description{
The function tests differentially expressed genes using generalized linear model likelihood-ratio tests
Contrasts in the matrix are sequentially fitted
}
\examples{
expression <- matrix(rnbinom(10000,mu=5,size=2),ncol=4)
design <- cbind(baseline=c(1,1,1,1), x=c(1,1,-1,-1), y=c(-1,1,-1,1))
edgeObj <- edgeBuild(expression, design)
edgeFit <- glmFit(edgeObj, design)
contrast1 <- c(0,1, 0)
contrast2 <- c(0,1, -1)
contrast3 <- cbind(C1=contrast1, C2=contrast2)
edgeTest1 <- edgeTest(edgeFit, contrast1)
edgeTest2 <- edgeTest(edgeFit, contrast2)
edgeTest3 <- edgeTest(edgeFit, contrast3)
stopifnot(identical(names(edgeTest3), c("C1", "C2")))
}
\seealso{
\code{\link{edgeTestContrast}} to test one contrast vector
}

