% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edgeTestContrast}
\alias{edgeTestContrast}
\title{Test differentially expressed genes using edgeR}
\usage{
edgeTestContrast(fit, contrast)
}
\arguments{
\item{fit}{An \code{DGEGLM} object produced by \code{glmFit}}

\item{contrast}{a vector of contrasts}
}
\value{
A data frame containing information about differential expression
}
\description{
The function tests differentially expressed genes using generalized linear model likelihood-ratio tests
A single contrast vector is allowed
}
\examples{
expression <- matrix(rnbinom(10000,mu=5,size=2),ncol=4)
design <- cbind(baseline=c(1,1,1,1), x=c(1,1,-1,-1))
edgeObj <- edgeBuild(expression, design)
edgeFit <- glmFit(edgeObj, design)
edgeFitTest <- edgeTestContrast(edgeFit, contrast=c(0,1))
}
\seealso{
\code{\link{glmFit}} to produce the input object, \code{\link{edgeTest}} to run the test for a contrast matrix.
}

