% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset2DGEList.R
\name{eset2DGEList}
\alias{eset2DGEList}
\title{Transform an EexpressionSet to a DGEList object}
\usage{
eset2DGEList(eset, groupVar = "group")
}
\arguments{
\item{eset}{An \code{ExpressionSet} object}

\item{groupVar}{Character string, column in phenoDatta}
}
\value{
A \code{DGEList} object
}
\description{
Transform an EexpressionSet to a DGEList object
}
\examples{
eset <- new("ExpressionSet",
  exprs=matrix(rpois(120, 20), nrow=20,
               dimnames=list(LETTERS[1:20], letters[1:6])),
  phenoData = new("AnnotatedDataFrame", data.frame(Sample=letters[1:6],
                  group=gl(2, 3), row.names=letters[1:6])),
  featureData = new("AnnotatedDataFrame", data.frame(Feature=LETTERS[1:20],
                  row.names=LETTERS[1:20])))
eset2DGEList(eset)
}
