% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{estimateGLMDisp}
\alias{estimateGLMDisp}
\alias{estimateGLMDisp,EdgeObject-method}
\title{Estimate generalized linear model dispersion}
\usage{
estimateGLMDisp(object)

\S4method{estimateGLMDisp}{EdgeObject}(object)
}
\arguments{
\item{object}{An \code{EdgeObject} object}
}
\value{
An updated \code{EdgeObject} object
}
\description{
Estimate generalized linear model dispersion
}
\section{Methods (by class)}{
\itemize{
\item \code{EdgeObject}: Method for EdgeObject
}}

