% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeObject.R
\name{exampleEdgeObject}
\alias{exampleEdgeObject}
\title{Return an example of EdgeObject}
\usage{
exampleEdgeObject(nfeat = 20, nsample = 6, ngroup = 3, lambda = 10)
}
\arguments{
\item{nfeat}{Integer, number of features}

\item{nsample}{Integer, number of samples}

\item{ngroup}{Integer, number of groups, must be divisible by \code{nsample}.}

\item{lambda}{Integer, passed to \code{rpois} to generate random counts.}
}
\value{
An \code{EdgeObject}
}
\description{
Return an example of EdgeObject
}
\examples{
set.seed(1887) ## fix random generators
exampleEdgeObject()
exampleEdgeObject(50, 12, ngroup=4)
}
