% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportEdgeResult.R
\name{exportEdgeResult}
\alias{exportEdgeResult}
\title{Export dgeTest results}
\usage{
exportEdgeResult(
  edgeResult,
  outRootDir,
  action = c("ask", "append", "overwrite", "no")
)
}
\arguments{
\item{edgeResult}{A \code{EdgeResult} object}

\item{outRootDir}{Character string, output directory}

\item{action}{Character string, what happens if the output directory exists}
}
\value{
Invisibly returns \code{NULL}. Called for its side effects of writing
  result files to disk.
}
\description{
Export dgeTest results
}
