% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticGeneLevelPlots.R
\name{exportStaticGeneLevelPlots}
\alias{exportStaticGeneLevelPlots}
\title{Export static gene-level plots in PDF}
\usage{
exportStaticGeneLevelPlots(edgeResult, file)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{file}{Character string, the PDF file name}
}
\value{
Called for its side effect of writing plots to a PDF file; returns
  invisibly NULL.
}
\description{
Export static gene-level plots in PDF
}
