% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{extractTargetAnno}
\alias{extractTargetAnno}
\title{Extract AmpliSeq target annotation from a list of read count data.frames}
\usage{
extractTargetAnno(readCountsList)
}
\arguments{
\item{readCountsList}{A list of read count data.frames

The function is used internally to extract unique target annotation from
multiple runs}
}
\description{
Extract AmpliSeq target annotation from a list of read count data.frames
}
