% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{fData,DGEList-method}
\alias{fData,DGEList-method}
\title{Get fData}
\usage{
\S4method{fData}{DGEList}(object)
}
\arguments{
\item{object}{A DGEList}
}
\value{
A \code{data.frame} of feature (gene) annotations.
}
\description{
Get fData
}
