% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{fData,EdgeObject-method}
\alias{fData,EdgeObject-method}
\title{Get fData}
\usage{
\S4method{fData}{EdgeObject}(object)
}
\arguments{
\item{object}{An EdgeObject}
}
\value{
A \code{data.frame} of feature (gene) annotations.
}
\description{
Get fData
}
