% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{fData<-,DGEList,data.frame-method}
\alias{fData<-,DGEList,data.frame-method}
\title{Set fData}
\usage{
\S4method{fData}{DGEList,data.frame}(object) <- value
}
\arguments{
\item{object}{A DGEList}

\item{value}{A \code{data.frame}}
}
\value{
The updated \code{DGEList} object with new feature annotations.
}
\description{
Set fData
}
