% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FeatAnnoExprs-class}
\alias{FeatAnnoExprs-class}
\title{A class that contain feature annotation and expression matrix}
\arguments{
\item{exprs}{A matrix of expression}

\item{genes}{A data.frame}
}
\description{
A class that contain feature annotation and expression matrix
}
