% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{featureNames,EdgeObject-method}
\alias{featureNames,EdgeObject-method}
\title{Feature names}
\usage{
\S4method{featureNames}{EdgeObject}(object)
}
\arguments{
\item{object}{An EdgeObject}
}
\value{
A character vector of feature names.
}
\description{
Feature names
}
