% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{filterByCPM.DGEList}
\alias{filterByCPM.DGEList}
\title{Filter lowly expressed genes by CPM in DGEList}
\usage{
\method{filterByCPM}{DGEList}(obj, lib.size = NULL, minCPM = 1,
  minCount = minGroupCount(obj))
}
\arguments{
\item{obj}{A \code{DGEList} object}

\item{minCPM}{Numeric, the minimum CPM accepted as expressed in one sample}

\item{minCount}{Integer, how many samples must have CPM larger than \code{minCPM} to keep this gene?}
}
\value{
Another \code{DGEList} object, with lowly expressed genes removed. The original counts and gene annotation can be found in \code{counts.unfiltered} and \code{genes.unfiltered} fields, respectively. The logical vector of the filter is saved in the \code{cpmFilter} field.
}
\description{
Filter lowly expressed genes by CPM in DGEList
}
\examples{
set.seed(1887)
mat <- rbind(matrix(rbinom(150, 5, 0.25), nrow=25), rep(0, 6))
d <- DGEList(mat, group=rep(1:3, each=2), genes=data.frame(Gene=sprintf("Gene\%d", 1:nrow(mat))))
df <- filterByCPM(d)

nrow(df$counts.unfiltered) ## 26
nrow(df$counts) ## 25
}
