% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{filterByCPM.matrix}
\alias{filterByCPM.matrix}
\title{Filter lowly expressed genes by CPM}
\usage{
\method{filterByCPM}{matrix}(obj, minCPM = 1, minCount = 1)
}
\arguments{
\item{obj}{A matrix}

\item{minCPM}{Numeric, the minimum CPM accepted as expressed in one sample}

\item{minCount}{Integer, how many samples must have CPM larger than \code{minCPM} to keep this gene?}
}
\value{
A logical vector of the same length as the row count of the matrix. \code{TRUE} means the gene is reasonably expressed, and \code{FALSE} means the gene is lowly expressed and should be filtered (removed)
}
\description{
Filter lowly expressed genes by CPM
}
\examples{
set.seed(1887)
mat <- rbind(matrix(rbinom(125, 5, 0.25), nrow=25), rep(0, 5))
filterByCPM(mat)
}
