% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByCPM.R
\name{filterByCPM}
\alias{filterByCPM}
\title{Filter lowly expressed genes by counts per million (CPM)}
\usage{
filterByCPM(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{Other parameters}
}
\value{
The return value depends on the class of \code{obj}. See method-specific documentation.
}
\description{
Filter lowly expressed genes by counts per million (CPM)
}
