% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{fitGLM}
\alias{fitGLM}
\alias{fitGLM,EdgeObject-method}
\title{Fit generalized linear model}
\usage{
fitGLM(object, ...)

\S4method{fitGLM}{EdgeObject}(object, ...)
}
\arguments{
\item{object}{An \code{EdgeObject} object}

\item{...}{Passed to \code{\link[edgeR]{glmFit}}}
}
\value{
A \code{DGEGLM} object containing the GLM fit.

The fit object
}
\description{
Fit generalized linear model
}
\section{Methods (by class)}{
\itemize{
\item \code{fitGLM(EdgeObject)}: Method for EdgeObject

}}
