% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{gctFilename}
\alias{gctFilename}
\title{Get GCT filename from a directory}
\usage{
gctFilename(dir)
}
\arguments{
\item{dir}{Character string, path to a directory where a GCT file is saved}
}
\value{
Character string, full name of the GCT file
If no file is found, the function reports an error. If more than one file is
found, a warning message is raised, and only the first file is used.
}
\description{
Get GCT filename from a directory
}
