% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{geneIdentifierTypes}
\alias{geneIdentifierTypes}
\title{Return gene identifier types}
\usage{
geneIdentifierTypes(dgeResult)
}
\arguments{
\item{dgeResult}{An DgeResult object}
}
\value{
A character string indicating the gene identifiers found
The following terms are recognized: GeneID, EnsemblID, GeneSymbol, FeatureName
}
\description{
Return gene identifier types
}
\note{
Note that the order matters: FeatureName > GeneID = EnsemblID > GeneSymbol
}
