% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticGeneLevelPlots.R
\name{groupCol}
\alias{groupCol}
\title{Get automatic group color}
\usage{
groupCol(edgeObj, panel = "Set1")
}
\arguments{
\item{panel}{passed to \code{fcbrewer}}

\item{edgeR}{An EdgeObject or EdgeResult object}
}
\value{
A fcbase object
}
\description{
Get automatic group color
}
