% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{groups,EdgeObject-method}
\alias{groups,EdgeObject-method}
\title{Get sample groups from an EdgeObject object}
\usage{
\S4method{groups}{EdgeObject}(object)
}
\arguments{
\item{object}{An \code{EdgeObject} object}
}
\value{
A factor of sample group assignments.
}
\description{
Get sample groups from an EdgeObject object
}
