% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockdown.R
\name{gtKdTable}
\alias{gtKdTable}
\title{Print a kdTable nicely with gt}
\usage{
gtKdTable(kdTable, feature_label = "GeneSymbol")
}
\arguments{
\item{kdTable}{a knockdown table (kdTable) returned by \code{kdTable}}

\item{feature_label}{The column which contains feature label, gene symbol by
default}
}
\value{
A \code{gt} table object
}
\description{
Print a kdTable nicely with gt
}
