% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{hasCommonDisp}
\alias{hasCommonDisp}
\alias{hasCommonDisp,DGEList-method}
\alias{hasCommonDisp,EdgeObject-method}
\title{Tells whether common dispersion has been set}
\usage{
hasCommonDisp(object)

\S4method{hasCommonDisp}{DGEList}(object)

\S4method{hasCommonDisp}{EdgeObject}(object)
}
\arguments{
\item{object}{An object}
}
\value{
Logical, whether the common dispersion has been set.
}
\description{
Tells whether common dispersion has been set
}
\section{Methods (by class)}{
\itemize{
\item \code{hasCommonDisp(DGEList)}: Method for DGEList

\item \code{hasCommonDisp(EdgeObject)}: Method for EdgeObject

}}
