% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/outdated-Rd.R
\docType{methods}
\name{inferSV,DGEList,formula-method}
\alias{inferSV,DGEList,formula-method}
\title{Infer surrogate variable from DGEList object and design matrix}
\usage{
\S4method{inferSV}{DGEList,formula}(object, design, ...)
}
\arguments{
\item{object}{A DGEList object, \code{voom} will be used to transform the data}

\item{design}{Formula}

\item{object}{A DGEList object, \code{voom} will be used to transform the
data}

\item{design}{Formula}
}
\value{
Surrogate variable matrix

Surrogate variable matrix
}
\description{
Infer surrogate variable from DGEList object and design matrix

Infer surrogate variable from DGEList object and design matrix
}
