% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{inferSV,matrix,matrix-method}
\alias{inferSV,matrix,matrix-method}
\title{Infer surrogate variable from expression matrix and design matrix}
\usage{
\S4method{inferSV}{matrix,matrix}(object, design, ...)
}
\arguments{
\item{object}{An expression matrix}

\item{design}{Design matrix}
}
\value{
Surrogate variable matrix
}
\description{
Infer surrogate variable from expression matrix and design matrix
}
