% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{inferSV}
\alias{inferSV}
\alias{inferSV,matrix,matrix-method}
\alias{inferSV,DGEList,matrix-method}
\alias{inferSV,DGEList,formula-method}
\title{Infer surrogate variables}
\usage{
inferSV(object, design, ...)

\S4method{inferSV}{matrix,matrix}(object, design, ...)

\S4method{inferSV}{DGEList,matrix}(object, design, ...)

\S4method{inferSV}{DGEList,formula}(object, design, ...)
}
\arguments{
\item{object}{An object}

\item{design}{Design matrix or formula}

\item{...}{Other parameters}
}
\value{
A matrix of surrogate variables.
}
\description{
Infer surrogate variables
}
\section{Methods (by class)}{
\itemize{
\item \code{inferSV(object = matrix, design = matrix)}: method for matrix as input

\item \code{inferSV(object = DGEList, design = matrix)}: method for voom-transformed DGEList and design matrix 
as input

\item \code{inferSV(object = DGEList, design = formula)}: method for voom-transformed DGEList and formula
as input

}}
