% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{isEmptySV}
\alias{isEmptySV}
\title{Is the Surrogate Variable (SV) matrix empty}
\usage{
isEmptySV(sv)
}
\arguments{
\item{sv}{A surrogate variable (SV) matrix returned by \code{sva}}
}
\value{
\code{TRUE} if no valid SV was estimated; otherwise \code{FALSE}.
}
\description{
Is the Surrogate Variable (SV) matrix empty
}
\examples{
isEmptySV(matrix(0, 1,1))
isEmptySV(matrix(rnorm(5), nrow=5))
}
