% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{isSig}
\alias{isSig}
\alias{isSigPos}
\alias{isSigNeg}
\title{Return logical vector indicating which genes are significantly regulated}
\usage{
isSig(data.frame, sigFilter)

isSigPos(data.frame, sigFilter)

isSigNeg(data.frame, sigFilter)
}
\arguments{
\item{data.frame}{A \code{data.frame} that must pass \code{assertEdgeToptable}}

\item{sigFilter}{An SigFilter object}
}
\value{
A logical vector of the same length as the row number of the input data.frame
}
\description{
Return logical vector indicating which genes are significantly regulated
}
\section{Functions}{
\itemize{
\item \code{isSigPos()}: Returns which genes are significantly positively regulated

\item \code{isSigNeg()}: Returns which genes are significantly negatively regulated

}}
