% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{isUnsetPosLogFC}
\alias{isUnsetPosLogFC}
\alias{isUnsetNegLogFC}
\alias{isUnsetAveExpr}
\alias{isUnsetLR}
\alias{isUnsetPValue}
\alias{isUnsetFDR}
\title{Tells whether the threshold was not set}
\usage{
isUnsetPosLogFC(edgeSigFilter)

isUnsetNegLogFC(edgeSigFilter)

isUnsetAveExpr(edgeSigFilter)

isUnsetLR(edgeSigFilter)

isUnsetPValue(edgeSigFilter)

isUnsetFDR(edgeSigFilter)
}
\arguments{
\item{edgeSigFilter}{An EdgeSigFilter object}
}
\value{
Logical, whether the thresholds are the default values
}
\description{
Tells whether the threshold was not set
}
