% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isUnsetSigFilter}
\alias{isUnsetSigFilter}
\title{Whether the SigFilter is the default one}
\usage{
isUnsetSigFilter(object)
}
\arguments{
\item{object}{An SigFilter object}
}
\value{
Logical, whether it is unset
}
\description{
Whether the SigFilter is the default one
}
