% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockdown.R
\name{kdTable}
\alias{kdTable}
\title{Retrieve a knockdown table from edgeRes}
\usage{
kdTable(edgeRes, feature, feature_label = "GeneSymbol")
}
\arguments{
\item{edgeRes}{An \code{EdgeResult} object}

\item{feature}{The feature to be retrieved}

\item{feature_label}{The column which contains feature label, gene symbol by
default}
}
\value{
A compact table containing essential information about knockdown
}
\description{
Retrieve a knockdown table from edgeRes
}
