% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMethods.R
\name{likeHumanGeneSymbol}
\alias{likeHumanGeneSymbol}
\title{Guess wether a vector of strings look like human gene symbols}
\usage{
likeHumanGeneSymbol(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
Logical
\code{TRUE} is only returned if at least eighty percent features match the pattern
}
\description{
Guess wether a vector of strings look like human gene symbols
}
