% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LimmaSigFilter-class}
\alias{LimmaSigFilter-class}
\title{LimmaSigFilter Extending BaseSigFilter to filter genes base on aveExpr}
\description{
LimmaSigFilter Extending BaseSigFilter to filter genes base on aveExpr
}
\section{Slots}{

\describe{
\item{\code{aveExpr}}{Numeric, AveExpr threshold (larger values are kept)}
}}

