% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LimmaVoomResult-class}
\alias{LimmaVoomResult-class}
\title{The LimmaVoom Object that contains test results, dgeTable, and SigFilter}
\description{
The LimmaVoom Object that contains test results, dgeTable, and SigFilter
}
\section{Slots}{

\describe{
\item{\code{marrayLM}}{A \code{MArrayLM} class object that contains results of eBayesFit}

\item{\code{voom}}{The \code{voom} object}
}}

