% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{LimmaVoomResult}
\alias{LimmaVoomResult}
\title{Construct a LimmaVoomResult object}
\usage{
LimmaVoomResult(edgeObj, voom, marrayLM, dgeTables)
}
\arguments{
\item{edgeObj}{An EdgeObject.}

\item{voom}{The voom (\code{EList}) object.}

\item{marrayLM}{A MArrayLM object.}

\item{dgeTables}{A list of DGEtables.}
}
\value{
An \code{LimmaVoomResult} object.
}
\description{
Construct a LimmaVoomResult object
}
