% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{logCPM}
\alias{logCPM}
\title{Get settings in the EdgeSigFilter}
\usage{
logCPM(edgeSigFilter)
}
\arguments{
\item{edgeSigFilter}{An \code{EdgeSigFilter} object}
}
\value{
Numeric values of the logCPM filter
}
\description{
Get settings in the EdgeSigFilter
}
