% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFCmatrix.R
\name{logFCmatrix}
\alias{logFCmatrix}
\title{Extract a matrix of log2(fold-change) values}
\usage{
logFCmatrix(
  edgeResult,
  featureIdentifier = "GeneSymbol",
  contrasts = NULL,
  removeNAfeatures = TRUE,
  minAveExpr = NULL
)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{featureIdentifier}{Character, column name in \code{dgeTable} that will
be used as rownames of the result matrix}

\item{contrasts}{\code{NULL} or characters; if not \code{NULL}, only logFC
values of given contrasts will be returned}

\item{removeNAfeatures}{Logical, if \code{TRUE}, features containing \code{NA}
values are removed.}

\item{minAveExpr}{\code{NULL} or numeric. If set, features with aveExpr lower
than the given value is not considered. This option is helpful to remove
genes that are lowly expressed which yet show strong differential
expression.}
}
\value{
A numeric matrix of log2(fold-change) values with features in rows
  and contrasts in columns.
}
\description{
Extract a matrix of log2(fold-change) values
}
\note{
TODO: add edgeResult data example
}
