% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logFCpca.R
\name{logFCmatrixPCA}
\alias{logFCmatrixPCA}
\title{Perform principal component analysis to the log fold-change matrix}
\usage{
logFCmatrixPCA(lfc)
}
\arguments{
\item{lfc}{A matrix of log2 fold changes, with features in rows and
contrasts in columns}
}
\value{
A \code{PCAScoreMatrix} object

The function performs principal component analysis (PCA) to the log 
fold-change matrix. 

By using a column of zeros during the PCA analysis,
which was removed from the final result, the point of origin represents 
an ideal contrast which yield absolutely no differential gene expression. 
It is easier to interpret the PCA results with this transformation.
}
\description{
Perform principal component analysis to the log fold-change matrix
}
\examples{
my_lfc_mat <- matrix(rnorm(1000), nrow=100, ncol=10)
my_lfc_pca <- logFCmatrixPCA(my_lfc_mat)
my_lfc_pca
}
