% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitExpression.R
\name{matchKeyValue}
\alias{matchKeyValue}
\title{Internal function used by readBiokitExpression}
\usage{
matchKeyValue(keys, dfList, keyColumn, valueColumns)
}
\arguments{
\item{keys}{A vector of character strings}

\item{dfList}{A list of data.frames to be matched}

\item{keyColumn}{The column of keys in the data.frames of dfList}

\item{valueColumns}{One or more character strings, column names in the data.frames of dfList

#@importFrom ribiosUtils matchColumn}
}
\value{
A list of matrices, each containing one value column
}
\description{
Internal function used by readBiokitExpression
}

