% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{minGroupCount.DGEList}
\alias{minGroupCount.DGEList}
\title{Return the size of the smallest group defined in the \code{DGEList} object}
\usage{
\method{minGroupCount}{DGEList}(dgeList)
}
\arguments{
\item{dgeList}{A \code{DGEList} object}
}
\value{
Integer
}
\description{
Return the size of the smallest group defined in the \code{DGEList} object
}
\examples{
y <- matrix(rnbinom(12000,mu=10,size=2),ncol=6)
d <- DGEList(counts=y, group=rep(1:3,each=2))
minGroupCount(d) ## 2 
d2 <- DGEList(counts=y, group=rep(1:2,each=3))
minGroupCount(d2) ## 3
d3 <- DGEList(counts=y, group=rep(1:3, 1:3))
minGroupCount(d3) ## 1
}
