% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{minGroupCount.EdgeObject}
\alias{minGroupCount.EdgeObject}
\title{Return the size of the smallest group defined in the \code{EdgeObject} object}
\usage{
\method{minGroupCount}{EdgeObject}(edgeObj)
}
\description{
Return the size of the smallest group defined in the \code{EdgeObject} object
}
