% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{model.DGEList}
\alias{model.DGEList}
\title{Build design matrix from a DGEList object}
\usage{
model.DGEList(object, formula, ...)
}
\arguments{
\item{object}{A DGEList object}

\item{formula}{Formula, passed to \code{\link{model.matrix}}

Sample annotation is used to construct the formula}

\item{...}{Not used so far}
}
\value{
A design matrix.
}
\description{
Build design matrix from a DGEList object
}
