% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{modLogCPM}
\alias{modLogCPM}
\alias{modLogCPM,DGEList-method}
\alias{modLogCPM,EdgeObject-method}
\title{Modulated logCPM}
\usage{
modLogCPM(object, ...)

\S4method{modLogCPM}{DGEList}(object, prior.count = 2)

\S4method{modLogCPM}{EdgeObject}(object)
}
\arguments{
\item{object}{A DGEList object}

\item{...}{Other parameters}

\item{prior.count}{Integer, prior count.}
}
\value{
A numeric matrix of modulated log-CPM values.
}
\description{
Modulated logCPM
}
\section{Methods (by class)}{
\itemize{
\item \code{modLogCPM(DGEList)}: Method for DGEList

\item \code{modLogCPM(EdgeObject)}: Method for EdgeObject

}}
