% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{myGage}
\alias{myGage}
\title{Wrap the gage::gage method to report consistent results as the CAMERA method}
\usage{
myGage(logFC, gmtList, ...)
}
\arguments{
\item{logFC}{A named vector of logFC values of genes}

\item{gmtList}{A \code{\link[BioQC]{GmtList}} object containing gene-sets}

\item{...}{Other parameters passed to \code{\link[gage]{gage}}}
}
\description{
Wrap the gage::gage method to report consistent results as the CAMERA method
}
