% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{naOrSqrt}
\alias{naOrSqrt}
\title{Return either NA (if input is NULL) or sqrt}
\usage{
naOrSqrt(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
\code{NA} or sqrt value
}
\description{
Return either NA (if input is NULL) or sqrt
}
