% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{ncol,EdgeResult-method}
\alias{ncol,EdgeResult-method}
\title{Return number of samples}
\usage{
\S4method{ncol}{EdgeResult}(x)
}
\arguments{
\item{x}{An EdgeResults object}
}
\value{
Integer
}
\description{
Return number of samples
}
