% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{normBoxplot}
\alias{normBoxplot}
\title{Plot distribution of normalized counts}
\usage{
normBoxplot(before.norm, after.norm, ...)
}
\arguments{
\item{before.norm}{An \code{EdgeObject} before normalization.}

\item{after.norm}{An \code{EdgeObject} after normalization.}

\item{...}{Other parameters passed to \code{boxplot}.}
}
\value{
Called for its side effect of plotting; returns invisibly NULL.
}
\description{
Plot distribution of normalized counts
}
