% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{nrow,EdgeResult-method}
\alias{nrow,EdgeResult-method}
\title{Return number of features}
\usage{
\S4method{nrow}{EdgeResult}(x)
}
\arguments{
\item{x}{An EdgeResults object}
}
\value{
Integer
}
\description{
Return number of features
}
