% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{parseCameraContributingGenes}
\alias{parseCameraContributingGenes}
\title{Parse contributing genes by genesets from the result data.frame of the \code{CAMERA} method}
\usage{
parseCameraContributingGenes(cameraResTbl, genesets)
}
\arguments{
\item{cameraResTbl}{A \code{tibble} or \code{data.frame} holding results of \code{CAMERA}}

\item{genesets}{Character strings, geneset labels}
}
\value{
A list of gene symbols, indexed by geneset names that are found in the results.
}
\description{
Parse contributing genes by genesets from the result data.frame of the \code{CAMERA} method
}
