% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{parseGenesetsContributingGenes}
\alias{parseGenesetsContributingGenes}
\title{Parse contributing genes by genesets}
\usage{
parseGenesetsContributingGenes(str, genesets)
}
\arguments{
\item{str}{Character strings, containing contributing genes}

\item{genesets}{Character strings, geneset labels. Its length must match the length of \code{str}}
}
\value{
A \code{data.frame} containing genesets, genes, and statistics
}
\description{
Parse contributing genes by genesets
}
\examples{
parseGenesetsContributingGenes("AKR1C4(-1.25), AKR1D1(-1.11)", "Metabolism")
parseGenesetsContributingGenes(c("AKR1C4(-1.25), AKR1D1(-1.11)",
                         "AKT1(1.24), AKT2(1.11), AKT3(1.05)"),
                         c("Metabolism", "AKTs"))
}
