% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{parseMolPhenFeat}
\alias{parseMolPhenFeat}
\title{Parse feature information from molecular-phenotyping GCT files}
\usage{
parseMolPhenFeat(gctMatrix)
}
\arguments{
\item{gctMatrix}{A \code{GctMatrix} capturing the counts of 
a molecular phenotyping experiment}
}
\value{
A data.frame with following columns: \code{GeneID} (as integer), 
  \code{GeneSymbol} (as character), and \code{Transcript},
   with the original names as row names
}
\description{
Parse feature information from molecular-phenotyping GCT files
}
\seealso{
\code{\link{readMolPhenCoverageGct}}, which calls this function
  internally to parse molecular phenotyping gene features
}
