% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{pData,DGEList-method}
\alias{pData,DGEList-method}
\title{Get pData (sample annotation)}
\usage{
\S4method{pData}{DGEList}(object)
}
\arguments{
\item{object}{A DGEList}
}
\value{
A \code{data.frame} of sample annotations.
}
\description{
Get pData (sample annotation)
}
