% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{pData,EdgeObject-method}
\alias{pData,EdgeObject-method}
\title{Get pData}
\usage{
\S4method{pData}{EdgeObject}(object)
}
\arguments{
\item{object}{An EdgeObject}
}
\value{
A \code{data.frame} of sample annotations.
}
\description{
Get pData
}
