% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{pData<-,DGEList,data.frame-method}
\alias{pData<-,DGEList,data.frame-method}
\title{Set pData (sample annotation)}
\usage{
\S4method{pData}{DGEList,data.frame}(object) <- value
}
\arguments{
\item{object}{A DGEList}

\item{value}{A \code{data.frame}}
}
\value{
The updated \code{DGEList} object with new sample annotations.
}
\description{
Set pData (sample annotation)
}
