% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{pData<-,EdgeObject,data.frame-method}
\alias{pData<-,EdgeObject,data.frame-method}
\title{Set pData (sample annotation)}
\usage{
\S4method{pData}{EdgeObject,data.frame}(object) <- value
}
\arguments{
\item{object}{A DGEList}

\item{value}{A \code{data.frame}}
}
\value{
The updated \code{EdgeObject} with new sample annotations.
}
\description{
Set pData (sample annotation)
}
