% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockdown.R
\name{plotGeneKnockdown}
\alias{plotGeneKnockdown}
\title{Plot Gene knockdown with knockdown efficiency}
\usage{
plotGeneKnockdown(goiExpr)
}
\arguments{
\item{goiExpr}{A data.frame containing expression of gene of interest, which
must contain following columns: \code{log2_CPM}, \code{GROUP} (a column or
character strings, among which a case-sensitive \code{Control} is available)}
}
\description{
Plot Gene knockdown with knockdown efficiency
}
\note{
The function is just a prototype and will be likely updated to make it more
standardized and flexible.
}
