% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockdown.R
\name{plotKnockdown}
\alias{plotKnockdown}
\title{Plot gene expression with knockdown efficiency}
\usage{
plotKnockdown(
  goiExpr,
  exprsVar = "exprs",
  groupVar = "group",
  controlGroup = NULL,
  trans = "identity",
  exprsUnit = "Arbitrary Unit",
  test = c("wilcox.test", "t.test")
)
}
\arguments{
\item{goiExpr}{A data.frame containing expression of gene of interest in linear scale, which
must contain columns given below as \code{exprsVar} and \code{groupVar}.}

\item{exprsVar}{Character, the variable name of expression. The unit must be in linear scale, not in logarithmic scale, otherwise the knockdown efficiency calculation will be wrong.}

\item{groupVar}{Character, the variable name of grouping. The column must be a factor or a character.}

\item{controlGroup}{\code{NULL} or character. If \code{groupVar} is a factor and \code{controlGroup} is \code{NULL}, then the first level is assumed to be the control. Otherwise, \code{controlGroup} must be in the group variable.}

\item{trans}{Character, transformation of the y-axis, commonly used values include \code{identity} (do not transform), \code{log10}, and \code{log2}.}

\item{exprsUnit}{Character, unit name of expression (TPM, CPM, RPKM, etc.)}

\item{test}{Character, statistical test, \code{wilcox.test} and \code{t.test} are supported.}
}
\value{
A \code{ggplot} object displaying boxplots of gene expression across
  groups with a secondary axis showing knockdown efficiency.
}
\description{
Plot gene expression with knockdown efficiency
}
\examples{

myData <- data.frame(group=gl(3,4),
 exprs=as.vector(sapply(c(100, 10, 1), function(x) rnorm(4, x))))
plotKnockdown(myData)

myData2 <-  data.frame(group=rep(c("Vehicle", "Dose1", "Dose2"), each=4),
 exprs=as.vector(sapply(c(100, 10, 1), function(x) rnorm(4, x))))
plotKnockdown(myData2, controlGroup="Vehicle")
}
