% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopSigGenes.R
\name{plotTopSigGenes}
\alias{plotTopSigGenes}
\title{Plot top significantly differentially expressed genes by contrast}
\usage{
plotTopSigGenes(
  countDgeResult,
  n = 5,
  nSigned = NULL,
  identifier = "GeneSymbol"
)
}
\arguments{
\item{countDgeResult}{A \code{CountDgeResult} object, for instance \code{EdgeResult} or \code{LimmaVoomDgeResult}.}

\item{n}{Integer, how many genes should be visualized per contrast.}

\item{nSigned}{\code{NULL} or integer, in the later case the top \code{nSigned} genes from positively and negatively regulated genes are shown per contrast, respectively.}

\item{identifier}{Character string, column name in \code{genes} annotation to be used to index and display the genes.}
}
\value{
A \code{ggplot} object
}
\description{
Plot top significantly differentially expressed genes by contrast
}
\examples{
edgeObj <- exampleEdgeObject()
edgeRes <- dgeWithEdgeR(edgeObj)
plotTopSigGenes(edgeRes, n=6)
## display top two positive and top three negative genes
plotTopSigGenes(edgeRes, nSigned=2)
}
\seealso{
\code{\link{plotTopSigGenesByContrast}} plots one contrast at a time.
}
