% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopSigGenes.R
\name{plotTopSigGenesByContrast}
\alias{plotTopSigGenesByContrast}
\title{Plot top significantly differentially expressed genes by contrast}
\usage{
plotTopSigGenesByContrast(
  countDgeResult,
  contrast,
  n = 5,
  nSigned = NULL,
  identifier = "GeneSymbol"
)
}
\arguments{
\item{countDgeResult}{A \code{CountDgeResult} object, for instance \code{EdgeResult} or \code{LimmaVoomDgeResult}.}

\item{contrast}{A character string, or an index (integer), or a logical vector with one \code{TRUE} element, to indicate which contrast to plot}

\item{n}{Integer, how many genes should be visualized.}

\item{nSigned}{\code{NULL} or integer, in the later case the top \code{nSigned} genes from positively and negatively regulated genes are shown, respectively.}

\item{identifier}{Character string, column name in \code{genes} annotation to be used to index and display the genes.}
}
\value{
A \code{ggplot} object. If \code{nSigned} is not \code{NULL}, genes are 
  plotted with colors: blue indicating down-regulated genes and red indicating up-regulated genes.
}
\description{
Plot top significantly differentially expressed genes by contrast
}
\examples{
edgeObj <- exampleEdgeObject()
edgeRes <- dgeWithEdgeR(edgeObj)
plotTopSigGenesByContrast(edgeRes, n=6, contrast=1)
plotTopSigGenesByContrast(edgeRes, n=6, contrast=2)
## display top three positive and top three negative genes
plotTopSigGenesByContrast(edgeRes, nSigned=3, contrast=1)
plotTopSigGenesByContrast(edgeRes, nSigned=4, contrast=2)
}
\seealso{
\code{\link{plotTopSigGenesByContrast}} plots all contrasts at once.
}
