% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prcomp.DGEListList}
\alias{prcomp.DGEListList}
\title{Run principal component analysis on a DGEListList object}
\usage{
\method{prcomp}{DGEListList}(x, ntop = NULL, fun = function(x) cpm(x, log = TRUE), ...)
}
\arguments{
\item{x}{A \code{DGEListList} object}

\item{ntop}{NULL or integer. If set, only \code{ntop} top-variable genes are
used}

\item{fun}{Function, used to transform count data into continuous data used
by PCA}

\item{...}{Not used.}
}
\value{
A list of \code{prcomp} objects.
}
\description{
Run principal component analysis on a DGEListList object
}
