% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prcompExprs}
\alias{prcompExprs}
\title{Principal component analysis of an expression matrix}
\usage{
prcompExprs(matrix, ntop = NULL, scale = FALSE, nbin = NULL)
}
\arguments{
\item{matrix}{Numeric matrix. Features in rows and samples in columns.}

\item{ntop}{Integer or NULL. If not \code{NULL}, only \code{ntop} genes with
the highest variance are used for the calculation.}

\item{scale}{Logical, whether variance of features should be scaled to 1.
Default \code{FALSE}, as recommended by Nguyen et al. (2019)}

\item{nbin}{Integer. Genes are divided into \code{nbin} bins by their average gene expression signal, and top variable genes (approximately \code{ntop/nbin}) are selected from each bin. If \code{NULL} or \code{NA}, an automatic value (100, or \code{nrow(matrix) \%/\% 10} when fewer are 1000 genes are used as input) is used. It is only used when \code{ntop} is not NULL.}
}
\value{
A \code{prcomp} object.
}
\description{
Principal component analysis of an expression matrix
}
\examples{

myTestExprs <- matrix(rnorm(1000), ncol=10, byrow=FALSE)
myTestExprs[1:50, 6:10] <- myTestExprs[1:50, 6:10] + 2
myTopPca <- prcompExprs(myTestExprs, ntop=50, nbin=5)

}
\references{
Nguyen, Lan Huong, and Susan Holmes. "Ten Quick Tips for
Effective Dimensionality Reduction." PLOS Computational Biology 15, no. 6
(2019): e1006907
}
\seealso{
\code{\link{topVarRowsByMeanBinning}}
}
