% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{quantileRange}
\alias{quantileRange}
\title{Return a range determined by the quantile of the data}
\usage{
quantileRange(x, outlier = 0.01, symmetric = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{outlier}{Quantile (lower and higher) threshold}

\item{symmetric}{Logical, whether the range must be symmetric around zero.}
}
\value{
A numeric vector of two (\code{c(low, high)}).
}
\description{
Return a range determined by the quantile of the data
}
