% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{read_illumina_sampleSheet_xls}
\alias{read_illumina_sampleSheet_xls}
\title{Read Illumina MolPhen sample sheet from XLS files}
\usage{
read_illumina_sampleSheet_xls(file)
}
\arguments{
\item{file}{A XLS/XLSX file containing in the first sheet
the sample sheet of a molecular phenotyping experiment}
}
\value{
A \code{data.frame} annotating the samples
}
\description{
Read Illumina MolPhen sample sheet from XLS files
}
