% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{readAmpliSeq}
\alias{readAmpliSeq}
\title{Read AmpliSeq results into an ExpressionSet object}
\usage{
readAmpliSeq(readCountFiles, barcodeSummaryFiles,
  runNames = names(readCountFileList))
}
\arguments{
\item{readCountFiles}{Character string vector, names of read count files}

\item{barcodeSummaryFiles}{Character string vector, names of barcode summary files}
}
\description{
Read AmpliSeq results into an ExpressionSet object
}
\examples{
countFiles <- system.file("extdata/AmpliSeq_files/",
  sprintf("ReadCountFiles-\%d.xls", 1:3), package="ribiosNGS")
barcodeFiles <- system.file("extdata/AmpliSeq_files/",
  sprintf("BarcodeSummaryFile-\%d.xls", 1:3), package="ribiosNGS")
exampleAmpliSeq <- readAmpliSeq(countFiles, barcodeFiles, 
  runNames=c("R1", "R2", "R3"))
}

