% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitAsDGEList.R
\name{readBiokitFeatureAnnotation}
\alias{readBiokitFeatureAnnotation}
\title{Read feature annotation from Biokit directory}
\usage{
readBiokitFeatureAnnotation(
  dir,
  anno = c("refseq", "ensembl", "gencode"),
  verbose = FALSE
)
}
\arguments{
\item{dir}{Character string, a Biokit output directory.}

\item{anno}{Character, indicating the annotation type.}

\item{verbose}{Logical}
}
\value{
A \code{data.frame} containing feature annotation, with feature IDs 
  as characters in rownames. The data frame contains following columns 
  depending on the \code{anno} parameter: 
  \enumerate{
    \item FeatureName, the primary key of feature name as characters
    \item GeneID (refseq only) or EnsemblID (ensembl only)
    \item GeneSymbol
    \item mean: mean length
    \item median: median length
    \item longest_isoform: longest isoform
    \item merged: total length of merged exons
  }
  
The function depends on the \code{refseq.annot.gz} (\code{ensembl.annot.gz}) 
and \code{refseq.geneLength.gz} (\code{ensembl.geneLength.gz}) files in the 
biokit directory.

If \code{.annot.gz} file is not found (which can be the case, for instance,
  when older biokit output directories are used), feature annotation is
  read from the count GCT file. The resulting \code{data.frame} will only 
  contain two columns: \code{FeatureName} and \code{Description}.
  
If \code{.geneLength.gz} file is not found, no gene length information is
 appended.
}
\description{
Read feature annotation from Biokit directory
}
\examples{
## TODO add small example files
}
