% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitAsDGEList.R
\name{readBiokitPhenodata}
\alias{readBiokitPhenodata}
\title{Read Biokit phenodata}
\usage{
readBiokitPhenodata(dir, verbose = FALSE)
}
\arguments{
\item{dir}{Character string, Biokit output directory}

\item{verbose}{Logical}
}
\value{
A \code{data.frame} with sample annotation in columns, and sample 
  names (identical as the names in gct files, character strings) are row 
  names. Nmes of the first three columns are fixed:
  \enumerate{
    \item \code{SampleName}, \code{SampleID} and \code{group} concatenated by 
       underscore
    \item \code{SampleID}
    \item \code{group}
  }
  
The function depends on the \code{annot/phenoData.meta} file in the biokit
 output directory.
}
\description{
Read Biokit phenodata
}
