% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{readCameraResults}
\alias{readCameraResults}
\title{Read CAMERA results into a tibble object}
\usage{
readCameraResults(file, minNGenes = 3, maxNGenes = 1000)
}
\arguments{
\item{file}{CAMERA results file}

\item{minNGenes}{NULL or integer, genesets with fewer genes are filtered out}

\item{maxNGenes}{NULL or integer, genesets with more genes are filtered out}
}
\description{
Read CAMERA results into a tibble object
}
