% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readMatrix}
\alias{readMatrix}
\title{Read in numeric matrix from tab-delimited format written by \code{writeMatrix}}
\usage{
readMatrix(file, row.names = TRUE)
}
\arguments{
\item{file}{file to be read in}

\item{row.names}{Logical, whether the first column contains row names (should be consistent with the settign in \code{writeMatrix})}
}
\value{
Matrix
}
\description{
\code{readMatrix} reads a matrix written by \code{writeMatrix} into a R session
}
\details{
Default behaviour of \code{read.table} is adapted to the convention used in \code{writeMatrix}
}
\examples{
test.mat <- matrix(rnorm(1000), nrow=10, dimnames=list(LETTERS[1:10], 1:100))
tmpfile <- tempfile()
writeMatrix(test.mat, tmpfile)
readin.mat <- readMatrix(tmpfile)
if(require(ribiosUtils)) identicalMatrix(test.mat, readin.mat)
}

