% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{readMolPhenAsDGEList}
\alias{readMolPhenAsDGEList}
\title{Read molecular phenotyping output folder into a DGEList object}
\usage{
readMolPhenAsDGEList(dir)
}
\arguments{
\item{dir}{Path of molecular phenotyping output folder, generated by the mpsnake tool.}
}
\value{
A \code{DGEList} object containing counts, gene and sample annotation.
}
\description{
Read molecular phenotyping output folder into a DGEList object
}
\examples{
#todo
}
