% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{readMolPhenCoverageGct}
\alias{readMolPhenCoverageGct}
\title{Read molecular phenotyping coverage file}
\usage{
readMolPhenCoverageGct(file)
}
\arguments{
\item{file}{Character string, a coverage GCT file of a molecular 
phenotyping experiment.}
}
\value{
A list of two elements: \code{coverage}, which represents the
   coverage matrix, and \code{genes}, which represents feature annotation.
}
\description{
Read molecular phenotyping coverage file
}
\examples{
mpsCov <- readMolPhenCoverageGct(system.file(file.path("extdata",
    "AmpliSeq_files",
    "MolPhen-coverage-example-20200115.gct"), 
  package="ribiosNGS")

}
