% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleAnnotationFileForEdgeR.R
\name{readSampleAnnotationFileForEdgeR}
\alias{readSampleAnnotationFileForEdgeR}
\title{Read sample annotation from tab-delimited file for EdgeR analysis}
\usage{
readSampleAnnotationFileForEdgeR(file = NULL, sampleNames = NULL)
}
\arguments{
\item{file}{Character string, a tab-delimited file. Row names must be present.
If \code{NULL}, \code{sampleNames} must be provided.}

\item{sampleNames}{Character string, giving sample names. If \code{NULL}, 
\code{file} must be provided.}
}
\value{
A \code{data.frame} containing sample annotation, removing
  `group`, `lib.size`, and `norm.factors` because they will be added
  by the edgeR pipeline
  
Either \code{file} or \code{sampleNames} must be provided. If both provided,
information in \code{file} has priority.
}
\description{
Read sample annotation from tab-delimited file for EdgeR analysis
}
\examples{
phenoDataFile <- system.file("extdata/phenoData/test-phenoData.txt",
  package="ribiosNGS")
readSampleAnnotationFileForEdgeR(phenoDataFile)
readSampleAnnotationFileForEdgeR(file=NULL, sampleNames=as.character(1:4))
}
