% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosNGS-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DGEList}
\alias{DesignContrast}
\alias{cpm}
\title{Objects exported from other packages}
\value{
A \code{DGEList} object.

A \code{DesignContrast} object.

A numeric matrix of counts per million values.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{edgeR}{\code{\link[edgeR:cpm]{cpm()}}, \code{\link[edgeR:DGEList]{DGEList()}}}

  \item{ribiosExpression}{\code{\link[ribiosExpression:DesignContrast]{DesignContrast()}}}
}}

