% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{removeCherryPickingRepeat}
\alias{removeCherryPickingRepeat}
\title{Remove samples that are repeated in cherry-picking runs}
\usage{
removeCherryPickingRepeat(eset, cherryPickingRun)
}
\arguments{
\item{eset}{An ExpressionSet object returned by \code{\link{readAmpliSeq}}}

\item{cherryPickingRun}{Character string(s), name(s) of cherry picking runs

#' @examples 
countFiles <- system.file("extdata/AmpliSeq_files/",
  sprintf("ReadCountFiles-%d.xls", 1:3), package="ribiosNGS")
barcodeFiles <- system.file("extdata/AmpliSeq_files/",
  sprintf("BarcodeSummaryFile-%d.xls", 1:3), package="ribiosNGS")
exampleAmpliSeq <- readAmpliSeq(coutFiles, barcodeFiles, 
  runNames=c("R1", "R2", "R3"))
newAmpliSeq <- removeCherryPickingRepeat(exampleAmpliSeq)}
}
\description{
Remove samples that are repeated in cherry-picking runs
}

