% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameContrast.R
\name{renameContrast}
\alias{renameContrast}
\title{Rename contrast by a pair of vectors}
\usage{
renameContrast(edgeResult, oldContrastName, newContrastName)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{oldContrastName}{A vector of character strings giving old contrast names}

\item{newContrastName}{completeA vector of character strings giving new contrast names, which match the \code{oldContrastName} one to one.}
}
\value{
A new \code{EdgeResult} object
}
\description{
Rename contrast by a pair of vectors
}
