% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameContrast.R
\name{renameContrastByFunc}
\alias{renameContrastByFunc}
\title{Rename contrast by a function}
\usage{
renameContrastByFunc(edgeResult, func)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{func}{A function receiving a vector of character strings as input, and returns
  another vector of the same length as output, for instance \code{gsub}.
The function can be called to rename contrasts}
}
\value{
A new \code{EdgeResult} object
}
\description{
Rename contrast by a function
}
