% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{replaceNAwithZero}
\alias{replaceNAwithZero}
\title{Replace NA counts with zero counts}
\usage{
replaceNAwithZero(edgeObj)
}
\arguments{
\item{edgeObj}{An EdgeObject object}
}
\value{
An EdgeObject object
}
\description{
Replace NA counts with zero counts
}
