% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosNGS.R
\docType{package}
\name{ribiosNGS-package}
\alias{ribiosNGS}
\alias{ribiosNGS-package}
\title{ribiosNGS: Utilities for next-generation sequencing data analysis in ribios}
\description{
NGS tasks in ribios, primarily differetial gene expression
    analysis.
}
\details{
ribiosNGS package provides functionalities used for next-generation sequencing data
}
\author{
\strong{Maintainer}: Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID}) [contributor]

}
\keyword{internal}
