\name{ribiosPlot-package}
\alias{ribiosPlot-package}
\alias{ribiosPlot}
\docType{package}
\title{
  ribiosPlot: Plotting in BIOS
}
\description{
  The package provides data structures and functions to plot graphics in
  BIOS with the help of R. It is a collection of code snippets I used to
  generate plots to visualize omic data. 
}
\details{
\tabular{ll}{
Package: \tab ribiosPlot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-09-21\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Jitao David Zhang <jitao_david.zhang@roche.com>
}
\keyword{ package }
