% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RiboSeq-class}
\alias{RiboSeq-class}
\title{A S4-class object representing a RiboSeq experiment}
\description{
A S4-class object representing a RiboSeq experiment
}
\section{Slots}{

\describe{
\item{\code{RNA}}{A \code{DGEList} object that represents RNA data}

\item{\code{RPF}}{A \code{DGEList} object that represents ribosome profiling 
(RPF) data}

\item{\code{groups}}{A factor vector that indicate sample groups}
}}

