% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{riboSeqAnalysisObject-class}
\alias{riboSeqAnalysisObject-class}
\title{A S4-class represnts analysis results of a RiboSeq experiment}
\description{
A S4-class represnts analysis results of a RiboSeq experiment
}
