\name{rowscale-methods}
\docType{methods}
\alias{rowscale-methods}
\alias{rowscale,matrix-method}
\title{ Row scale }
\description{
  Scaling by row (the \code{scale} function scales by column by default)
}
\section{Methods}{
\describe{

\item{\code{signature(object = "matrix")}}{Scale a matrix by row}
}}
\keyword{methods}
