\name{royalbluered}
\alias{royalbluered}
\alias{royalredblue}
\alias{turyeb}
\alias{redgreen}
\alias{greenred}
\alias{bluered}
\alias{redblue}
\title{Three-color panels}
\description{
  Precompiled three-color panels for visualization purposes
}
\usage{
royalbluered(n)
royalredblue(n)
turyeb(n)
redgreen(n)
greenred(n)
bluered(n)
redblue(n)
}
\arguments{
  \item{n}{Number of colors needed}
}
\details{
  See function definitions
}
\value{Character vector of length \code{n} coding colors}
\author{Jitao David Zhang}
\seealso{
  \code{blackyellow} for two-color systems
}
\examples{
display.threecolor.panels()
}

