% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{sampleNames,EdgeObject-method}
\alias{sampleNames,EdgeObject-method}
\title{Sample names}
\usage{
\S4method{sampleNames}{EdgeObject}(object)
}
\arguments{
\item{object}{An EdgeObject}
}
\value{
A character vector of sample names.
}
\description{
Sample names
}
