% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,DGEList-method}
\alias{show,DGEList-method}
\title{Show DGEList}
\usage{
\S4method{show}{DGEList}(object)
}
\arguments{
\item{object}{A DGEList object}
}
\value{
Called for its side effect of printing; returns invisibly NULL.
}
\description{
Show DGEList
}
