% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,DGEListList-method}
\alias{show,DGEListList-method}
\title{Show DGEListList}
\usage{
\S4method{show}{DGEListList}(object)
}
\arguments{
\item{object}{A DGEListList object}
}
\value{
Called for its side effect of printing; returns invisibly NULL.
}
\description{
Show DGEListList
}
