% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,EdgeResult-method}
\alias{show,EdgeResult-method}
\title{Show an EdgeResult object}
\usage{
\S4method{show}{EdgeResult}(object)
}
\arguments{
\item{object}{An EdgeResult object}
}
\value{
Invisibly returns the formatted message string.
}
\description{
Show an EdgeResult object
}
