% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,EdgeSigFilter-method}
\alias{show,EdgeSigFilter-method}
\title{Show an EdgeSigFilter object}
\usage{
\S4method{show}{EdgeSigFilter}(object)
}
\arguments{
\item{object}{An SigFilter object}
}
\value{
Invisibly returns the formatted message strings.
}
\description{
Show an EdgeSigFilter object
}
