% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,LimmaSigFilter-method}
\alias{show,LimmaSigFilter-method}
\title{Show an LimmaSigFilter object}
\usage{
\S4method{show}{LimmaSigFilter}(object)
}
\arguments{
\item{object}{An LimmaSigFilter object}
}
\value{
Invisibly returns the formatted message strings.
}
\description{
Show an LimmaSigFilter object
}
