% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SigFilter-class}
\alias{SigFilter-class}
\alias{EdgeSigFilter}
\alias{LimmaSigFilter}
\title{Base result filter for significantly regulated genes}
\usage{
EdgeSigFilter(logFC, posLogFC, negLogFC, pValue, FDR, logCPM)

LimmaSigFilter(logFC, posLogFC, negLogFC, pValue, FDR, aveExpr)
}
\description{
Base result filter for significantly regulated genes
}
\section{Functions}{
\itemize{
\item \code{EdgeSigFilter}: EdgeSigFilter

\item \code{LimmaSigFilter}: LimmaSigFilter
}}

\section{Slots}{

\describe{
\item{\code{posLogFC}}{Numeric, positive logFC threshold (larger values are kept)}

\item{\code{negLogFC}}{Numeric, negative logFC threshold (more negative values 
are kept)}

\item{\code{pValue}}{Numeric, p-value treshold (smaller values are kept)}

\item{\code{FDR}}{Numeric, FDR treshold}
}}

\examples{
esf <- EdgeSigFilter(logFC=2, FDR=0.05, logCPM=0)
LimmaSigFilter(logFC=1, FDR=0.05, aveExpr=10)
}
