% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigFilter}
\alias{sigFilter}
\title{Retrieve SigFilter objects from other objects
Return the SigFilter in use}
\usage{
sigFilter(countDgeResult)
}
\arguments{
\item{countDgeResult}{An \code{countDgeResult} object}
}
\value{
An \code{SigFilter} object
}
\description{
Retrieve SigFilter objects from other objects
Return the SigFilter in use
}
