% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticGeneLevelPlots.R
\name{sigGeneBarchart}
\alias{sigGeneBarchart}
\title{Barchart of significantly regulated genes}
\usage{
sigGeneBarchart(
  edgeResult,
  logy = FALSE,
  scales = list(x = list(rot = 45), y = list(alternating = 1, tck = c(1, 0))),
  stack = FALSE,
  ylab = "Significant DEGs",
  col = c(positive = "orange", negative = "lightblue"),
  auto.key = list(columns = 2),
  ...
)
}
\arguments{
\item{edgeResult}{An EdgeResult object}

\item{logy}{Logical, whether y-axis should be log-10 transformed}

\item{scales}{passed to \code{lattice::barchart}}

\item{stack}{passed to \code{lattice::barchart}}

\item{ylab}{passed to \code{lattice::barchart}}

\item{col}{passed to \code{lattice::barchart}}

\item{auto.key}{passed to \code{lattice::barchart}}

\item{...}{passed to \code{lattice::barchart}}
}
\value{
A \code{trellis} object (lattice barchart).
}
\description{
Barchart of significantly regulated genes
}
