% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigGeneCounts}
\alias{sigGeneCounts}
\title{Return counts of significantly regulated genes}
\usage{
sigGeneCounts(countDgeResult)
}
\arguments{
\item{countDgeResult}{An EdgeResult object}
}
\value{
A data.frame containing counts of positively and negatively regulated
   genes, the sum, as well as total number of features
}
\description{
Return counts of significantly regulated genes
}
