% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigGeneIdentifiers}
\alias{sigGeneIdentifiers}
\title{Return gene identifiers of significant DGEs}
\usage{
sigGeneIdentifiers(dgeResult, contrast, sigFunc = isSig)
}
\arguments{
\item{dgeResult}{An DgeResult object.}

\item{contrast}{A character string, a contrast of interest.}

\item{sigFunc}{A function, defining the type of significant genes.}
}
\value{
A vector of character strings indicating the gene identifiers that are significantly regulated. If no defined types are found, either rownames or the first column is returned
}
\description{
Return gene identifiers of significant DGEs
}
\seealso{
\code{\link{geneIdentifierTypes}}
}
