% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigGenes}
\alias{sigGenes}
\alias{sigPosGenes}
\alias{sigNegGenes}
\title{Return significantly regulated genes of all contrasts}
\usage{
sigGenes(countDgeResult, value = NULL)

sigPosGenes(countDgeResult, value = NULL)

sigNegGenes(countDgeResult, value = NULL)
}
\arguments{
\item{countDgeResult}{An EdgeResult object}

\item{value}{\code{NULL} or character string, if not \code{NULL}, it must be a column name in the feature annotation data.}
}
\value{
A list of vectors of identifiers
}
\description{
Return significantly regulated genes of all contrasts
}
\section{Functions}{
\itemize{
\item \code{sigPosGenes}: Only return significantly positively regulated genes

\item \code{sigNegGenes}: Only return significantly negatively regulated genes
}}

\note{
TODO fix: add InputFeature
}
