% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-slurm-lsf.R
\name{slurmEdgeR}
\alias{slurmEdgeR}
\title{Send an edgeR analysis job to SLURM}
\usage{
slurmEdgeR(
  dgeList,
  designContrast,
  outdir = "edgeR_output",
  outfilePrefix = "an-unnamed-project-",
  overwrite = c("ask", "overwrite", "append", "no"),
  mps = FALSE,
  limmaVoom = FALSE,
  appendGmt = NULL,
  qos = c("3h", "1d", "3d", "15d", "interactive", "preempt"),
  rootPath = "/apps/rocs/pRED/groups/bioinfo/geneexpression",
  debug = FALSE
)
}
\arguments{
\item{dgeList}{An \code{DGEList} object with \code{counts}, \code{genes},
and \code{samples}}

\item{designContrast}{The DesignContrast object to model the data}

\item{outdir}{Output directory of the edgeR script. Default value
"edgeR_output".}

\item{outfilePrefix}{Prefix of the output files. It can include directories,
e.g. \code{"data/outfile-"}. In case of \code{NULL}, temporary files will be
created.}

\item{overwrite}{If \code{ask}, the user is asked before an existing output
directory is overwritten. If \code{yes}, the job will start and an existing
directory will be overwritten anyway. If \code{no}, and if an output
directory is present, the job will not be started.}

\item{mps}{Logical, whether molecular-phenotyping analysis is run.}

\item{limmaVoom}{Logical, whether the limma-voom model is run instead of the edgeR model.}

\item{appendGmt}{\code{NULL} or character string, path to an additional GMT
file for gene-set analysis. The option is passed to
\code{\link{slurmEdgeRcommand}} and then to \code{\link{edgeRcommand}}.}

\item{qos}{Character, specifying Quality of Service of Slurm. Available values include \code{short} (recommended default, running time cannot exceed 3 hours), \code{interactive} (useful if you wish to get the results from an interactive session), and \code{long} (useful if the job is expected to run more than three hours.)
using \code{srun} and the 'interaction' queue of jobs instead of using
\code{sbatch}.}

\item{rootPath}{Character string, the directory of geneexpression scripts, under which \code{bin/ngsDge_edgeR.Rscript} is found.}

\item{debug}{Logical, if \code{TRUE}, the source code of Rscript is used instead of
the installed version. The option is passed to \code{edgeRcommand}.}
}
\value{
A list of two items, \code{command}, the command line call, and
\code{output}, the output of the SLURM command in bash
}
\description{
Send an edgeR analysis job to SLURM
}
\note{
Even if the output directory is empty, if \code{overwrite} is set to
\code{no} (or if the user answers \code{no}), the job will not be started.
}
\examples{

 mat <- matrix(rnbinom(100, mu=5, size=2), ncol=10)
 rownames(mat) <- sprintf("gene\%d", 1:nrow(mat))
 myFac <- gl(2,5, labels=c("Control", "Treatment"))
 y <- edgeR::DGEList(counts=mat, group=myFac)
 myDesign <- model.matrix(~myFac); colnames(myDesign) <- levels(myFac)
 myContrast <- limma::makeContrasts(Treatment, levels=myDesign)
 myDescon <- DesignContrast(myDesign, myContrast)
 ## \dontrun{
 ## slurmEdgeR(y, myDescon, outfilePrefix="test", outdir=tempdir())
 ## }

}
