% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/annotateMethods.R
\name{sniffFeatures}
\alias{sniffFeatures}
\alias{sniffFeatures,EdgeObject-method}
\title{Sniff features}
\usage{
sniffFeatures(object)

\S4method{sniffFeatures}{EdgeObject}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Sniff features
}
\section{Methods (by class)}{
\itemize{
\item \code{EdgeObject}: sniff features for EdgeObject
}}

