% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{split,DGEList,factor-method}
\alias{split,DGEList,factor-method}
\title{Split DGEList by a factor into a DGEList2 object}
\usage{
\S4method{split}{DGEList,factor}(x, f, drop = FALSE, keep.lib.sizes = TRUE, sampleDropLevels = TRUE)
}
\arguments{
\item{x}{A \code{DGEList} object}

\item{f}{A factor}

\item{drop}{Logical, whether unused levels in the factor should be dropped}

\item{keep.lib.sizes}{Logical, whether library sizes are kept}

\item{sampleDropLevels}{logical, whether unused levels of factors in the sample annotation data frame should be dropped}
}
\description{
Split DGEList by a factor into a DGEList2 object
}
