% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topDgeExpression.R
\name{topDgeExpression}
\alias{topDgeExpression}
\title{Return raw expression of top differentially expressed genes of multiple contrasts}
\usage{
topDgeExpression(
  edgeResult,
  ntop = 10,
  contrast = NULL,
  exprsFun = function(dgeList) cpm(dgeList, log = TRUE)
)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object.}

\item{ntop}{Integer, number of top differentially expressed genes.}

\item{contrast}{\code{NULL}, which means all contrasts are considered, or
a vector of character strings or of integer values, which specify
which contrasts are considered.}

\item{exprsFun}{A function to derive expression values from \code{EdgeResult}.}
}
\value{
A \code{tibble} object with \code{Contrast} in the first column and
   a wide table of raw expression, feature annotation, and sample annotations
   in the rest columns
}
\description{
Return raw expression of top differentially expressed genes of multiple contrasts
}
\seealso{
\code{\link{topDgeExpressionByContrast}}
}
