% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topDgeExpression.R
\name{topDgeExpressionByContrast}
\alias{topDgeExpressionByContrast}
\title{Return raw expression of top differentially expressed genes of one contrast}
\usage{
topDgeExpressionByContrast(
  edgeResult,
  ntop = 10,
  contrast,
  exprsFun = function(dgeList) cpm(dgeList, log = TRUE)
)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object.}

\item{ntop}{Integer, number of top differentially expressed genes.}

\item{contrast}{A character string or an integer value, specifying which
contrast is considered.}

\item{exprsFun}{A function to derive expression values from \code{EdgeResult}.}
}
\value{
A \code{tibble} object with \code{Contrast} in the first column and
   a wide table of raw expression, feature annotation, and sample annotations
   in the rest columns
}
\description{
Return raw expression of top differentially expressed genes of one contrast
}
