% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{topVarRowIndByMeanBinning}
\alias{topVarRowIndByMeanBinning}
\title{Bin rows by mean expression, and return integer indices of rows with the highest variations from each bin}
\usage{
topVarRowIndByMeanBinning(matrix, ntop = NULL, nbin = NULL)
}
\arguments{
\item{matrix}{A numeric matrix}

\item{ntop}{Integer, the total number of rows that are expected to be returned}

\item{nbin}{Integer, how many bins should be formed? If \code{NULL}, an automatic value is used.}
}
\value{
An integer vector of rows, containing indices

The function first bin rows by their average expression profile, and choose rows with the highest variation from each bin.
}
\description{
Bin rows by mean expression, and return integer indices of rows with the highest variations from each bin
}
\note{
In case \code{ntop} is not a multiple of \code{nbin}, the \code{modulo} genes with the loweset variation are removed to garantee that the resulting matrix has exactly \code{ntop} rows.
}
\examples{
# myMat <- matrix(rnorm(2000), ncol=10, byrow=FALSE)
# myTopVarMatInd <- topVarRowIndByMeanBinning(myMat, ntop=130, nbin=15)
# myTopVarMatInd2 <- topVarRowIndByMeanBinning(myMat, ntop=135, nbin=15)
}
