% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{topVarRowsByMeanBinning}
\alias{topVarRowsByMeanBinning}
\title{Bin rows by mean expression, and return rows with the highest variations from each bin}
\usage{
topVarRowsByMeanBinning(matrix, ntop = NULL, nbin = NULL)
}
\arguments{
\item{matrix}{A numeric matrix}

\item{ntop}{Integer, the total number of rows that are expected to be returned}

\item{nbin}{Integer, how many bins should be formed? If \code{NULL}, an automatic value is used.}
}
\value{
A subset of the matrix, containing exactly \code{ntop} rows.
}
\description{
Bin rows by mean expression, and return rows with the highest variations from each bin
}
\examples{
# myMat <- matrix(rnorm(2000), ncol=10, byrow=FALSE)
# myTopVarMat <- topVarRowsByMeanBinning(myMat, ntop=130, nbin=15)
# myTopVarMat2 <- topVarRowsByMeanBinning(myMat, ntop=135, nbin=15)
}
\seealso{
\code{\link{topVarRowIndByMeanBinning}}
}
