% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{posLogFC<-}
\alias{posLogFC<-}
\alias{negLogFC<-}
\alias{logFC<-}
\alias{pValue<-}
\alias{FDR<-}
\alias{logCPM<-}
\alias{aveExpr<-}
\alias{update.SigFilter}
\alias{`logFC<-`}
\alias{`posLogFC<-`}
\alias{`negLogFC<-`}
\alias{`aveExpr<-`}
\alias{`logCPM<-`}
\alias{`pValue<-`}
\alias{`FDR<-`}
\alias{update.EdgeSigFilter}
\alias{update.LimmaSigFilter}
\title{Update SigFilter}
\usage{
posLogFC(object) <- value

negLogFC(object) <- value

logFC(object) <- value

pValue(object) <- value

FDR(object) <- value

logCPM(object) <- value

aveExpr(object) <- value

\method{update}{SigFilter}(object, logFC, posLogFC, negLogFC, pValue, FDR, ...)

\method{update}{EdgeSigFilter}(object, logFC, posLogFC, negLogFC, pValue, FDR, logCPM, ...)

\method{update}{LimmaSigFilter}(object, logFC, posLogFC, negLogFC, pValue, FDR, aveExpr, ...)
}
\arguments{
\item{object}{An \code{SigFilter} object}

\item{value}{Numeric, vssigned threshold value}

\item{logFC}{Numeric, logFC filter value, optional.}

\item{posLogFC}{Numeric, positive logFC filter value, optional.}

\item{negLogFC}{Numeric, negative logFC filter value, optional.}

\item{pValue}{Numeric, pValue filter value, optional}

\item{FDR}{Numeric, FDR filter value, optional}

\item{...}{not used now}

\item{logCPM}{Numeric, logCPM filter value, optional (only for \code{EdgeSigFilter}).}

\item{aveExpr}{Numeric, aveExpr filter value, optional (only for \code{LimmaSigFilter}).}
}
\value{
An updated \code{SigFilter} object.

An updated \code{EdgeSigFilter} object.

An updated \code{LimmaSigFilter} object.
}
\description{
Update SigFilter
}
\section{Functions}{
\itemize{
\item \code{posLogFC(object) <- value}: Updates the posLogFC threshold value

\item \code{negLogFC(object) <- value}: Updates the negLogFC threshold value

\item \code{logFC(object) <- value}: Updates the posLogFC threshold value

\item \code{pValue(object) <- value}: Updates the pValue threshold value

\item \code{FDR(object) <- value}: Updates the FDR threshold value

\item \code{logCPM(object) <- value}: Updates the logCPM threshold value

\item \code{aveExpr(object) <- value}: Updates the aveExpr threshold value

}}
