% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{updateContrastMatrixWithSV}
\alias{updateContrastMatrixWithSV}
\title{Update a contrast matrix given a surrogate variable matrix}
\usage{
updateContrastMatrixWithSV(contrastMatrix, svMatrix)
}
\arguments{
\item{contrastMatrix}{A contrast matrix}

\item{svMatrix}{A surrogate-variable matrix, for instance returned by sva}
}
\value{
An updated matrix, with surrogate matrix variables appended to the rows
}
\description{
Update a contrast matrix given a surrogate variable matrix
}
\examples{
exCounts <- matrix(rpois(12000, 10), nrow=2000, ncol=6)
exCounts[1:100, 2:3] <- exCounts[1:100,2:3]+20
exDesign <- model.matrix(~gl(2,3))
colnames(exDesign) <- c("Baseline", "Treatment")
exContrast <- limma::makeContrasts("Treatment"="Treatment", levels=exDesign)
exVoomSvaRes <- voomSVA(exCounts, exDesign)
updateContrastMatrixWithSV(exContrast, exVoomSvaRes) 
}
