% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/outdated-Rd.R
\docType{methods}
\name{updateDesignMatrixBySVA,DGEList,formula-method}
\alias{updateDesignMatrixBySVA,DGEList,formula-method}
\title{Update design matrix by results of SVA}
\usage{
\S4method{updateDesignMatrixBySVA}{DGEList,formula}(object, design, ...)
}
\arguments{
\item{object}{A DGEList object}

\item{design}{Formula to infer design}

\item{object}{A DGEList object}

\item{design}{Formula to infer design}
}
\value{
Updated design matrix including surrogate variables

Updated design matrix including surrogate variables
}
\description{
Update design matrix by results of SVA

Update design matrix by results of SVA
}
