% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{updateDesignMatrixBySVA}
\alias{updateDesignMatrixBySVA}
\alias{updateDesignMatrixBySVA,DGEList,formula-method}
\title{Update design matrix by SVA}
\usage{
updateDesignMatrixBySVA(object, design, ...)

\S4method{updateDesignMatrixBySVA}{DGEList,formula}(object, design, ...)
}
\arguments{
\item{object}{An object}

\item{design}{Design matrix or formula}

\item{...}{Other parameters}
}
\value{
A design matrix updated with surrogate variables.
}
\description{
Update design matrix by SVA
}
\section{Methods (by class)}{
\itemize{
\item \code{updateDesignMatrixBySVA(object = DGEList, design = formula)}: for DGEList and formula

}}
