% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{updateSigFilter}
\alias{updateSigFilter}
\title{Update the EdgeSigFilter}
\usage{
updateSigFilter(edgeResult, logFC, posLogFC, negLogFC, logCPM, LR, pValue, FDR)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{logFC}{Numeric}

\item{posLogFC}{Numeric}

\item{negLogFC}{Numeric}

\item{logCPM}{Numeric}

\item{LR}{Numeric}

\item{pValue}{Numeric}

\item{FDR}{Numeric}
}
\value{
An updated \code{EdgeResult} object with updated \code{EdgeSigFilter}
}
\description{
Update the EdgeSigFilter
}
