\name{GSEA_DATA_DIR}
\alias{GSEA_DATA_DIR}
\alias{GSEA_ANNOTATION_DIR}
\alias{GSEA_GENESET_DIR}
\alias{JAVA_BIN}
\alias{GSEA_JAR}
\alias{DEFAULT_GMT}
\alias{DEFAULT_CHIP}
\docType{data}
\title{
  Variables useful for GSEA analysis in Roche bioinformatics
  environment.
}
\description{
  These variables ar useful for performing GSEA analysis in Roche
  bioinformatics environment.
}
