% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMethods.R
\name{volcanoPlot}
\alias{volcanoPlot}
\alias{volcanoPlot,EdgeResult-method}
\title{Volcano plot}
\usage{
volcanoPlot(object, ...)

\S4method{volcanoPlot}{EdgeResult}(
  object,
  contrast = NULL,
  freeRelation = FALSE,
  colramp = ribiosPlot::heat,
  multipage = FALSE,
  yValue = c("PValue", "FDR"),
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  topLabel = NULL,
  labelType = NULL,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}

\item{contrast}{Character, contrast of interest. If \code{NULL}, 
all contrasts are used}

\item{freeRelation}{Logical.}

\item{colramp}{Function, color palette.}

\item{multipage}{Logical.}

\item{yValue}{Character string, either \code{PValue} or \code{FDR}.}

\item{xlim}{NULL or a numeric vector of two}

\item{ylim}{NULL or a numeric vector of two.}

\item{main}{Character, title.}

\item{topLabel}{NULL or an integer number, number of top features to be labelled}

\item{labelType}{NULL or a character string, a column name in the feature annotation}
}
\value{
Called for its side effect of plotting; returns invisibly NULL.
}
\description{
Volcano plot
}
\section{Methods (by class)}{
\itemize{
\item \code{volcanoPlot(EdgeResult)}: Method for EdgeResult

}}
