% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voomLimma.R
\name{voomLimma}
\alias{voomLimma}
\title{Perform the voom+limma procedure}
\usage{
voomLimma(
  dgeList,
  design,
  contrasts,
  normalize.method = "none",
  block = NULL,
  correlation = NULL,
  weights = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{dgeList}{A DGEList object, it should be ideally already filtered}

\item{design}{The design matrix}

\item{contrasts}{The contrast matrix}

\item{normalize.method}{Character string, passed to \code{voom}, keep it \code{none} unless you are sure}

\item{block}{Blocking factor, passed to \code{voom}}

\item{correlation}{Correlation between duplicates, passed to \code{voom}}

\item{weights}{Weights, passed to \code{voom}}

\item{plot}{Logical, whether the variance-mean relationship should be ploted}

\item{...}{Passed to \code{\link[limma]{eBayes}}}
}
\value{
\code{MArrayLM} object returned by \code{\link[limma]{eBayes}}, with voom object in the \code{voom} element of the list
}
\description{
Perform the voom+limma procedure
}
\examples{
y <- matrix(rnbinom(10000,mu=5,size=2),ncol=4)
d <- edgeR::DGEList(counts=y, group=rep(1:2,each=2))
d <- edgeR::calcNormFactors(d)
design <- model.matrix(~gl(2,2))
colnames(design) <- c("baseline", "treatment")
contrasts <- limma::makeContrasts("treatment", levels=design)
dvl <- voomLimma(d, design=design, contrasts=contrasts)

}
