% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{voomSVA,DGEList,formula-method}
\alias{voomSVA,DGEList,formula-method}
\title{Detect surrogate variables from DGEList}
\usage{
\S4method{voomSVA}{DGEList,formula}(object, design)
}
\arguments{
\item{object}{A DGEList object}

\item{matrix}{Formula to infer design matrix}
}
\value{
A new DGEList object, including new items in the list: \code{voom}, \code{sv}, \code{designMatrix}, \code{designMatrixWithSV}, and \code{voomSVRemoved}.
}
\description{
Detect surrogate variables from DGEList
}
