% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{voomSVA,matrix,matrix-method}
\alias{voomSVA,matrix,matrix-method}
\title{Run SVA on a count matrix transformed by voom}
\usage{
\S4method{voomSVA}{matrix,matrix}(object, design)
}
\arguments{
\item{object}{A count matrix}

\item{design}{Design matrix}
}
\value{
SV matrix
}
\description{
Run SVA on a count matrix transformed by voom
}
\examples{
set.seed(1887)
exCounts <- matrix(rpois(12000, 10), nrow=2000, ncol=6)
exCounts[1:100, 2:3] <- exCounts[1:100,2:3]+20
exDesign <- model.matrix(~gl(2,3))
voomSVA(exCounts, design=exDesign)
}
