% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva.R
\name{voomSVA}
\alias{voomSVA}
\title{Apply SVA to voom-transformed count data}
\usage{
voomSVA(counts, designMatrix)
}
\arguments{
\item{counts}{A matrix of counts}

\item{designMatrix}{Design matrix}
}
\value{
The SV matrix
}
\description{
Apply SVA to voom-transformed count data
}
\examples{
set.seed(1887)
exCounts <- matrix(rpois(12000, 10), nrow=2000, ncol=6)
exCounts[1:100, 2:3] <- exCounts[1:100,2:3]+20
exDesign <- model.matrix(~gl(2,3))
voomSVA(exCounts, designMatrix=exDesign)
}
