% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBiokitSampleAnnotation.R
\name{writeBiokitSampleAnnotation}
\alias{writeBiokitSampleAnnotation}
\title{Write sample annotation into a tab-delimited file to start the Biokit
pipeline}
\usage{
writeBiokitSampleAnnotation(df, con)
}
\arguments{
\item{df}{A data.frame or anything that can be converted to a data.frame}

\item{con}{Connection, can be a character string indicating file name}
}
\value{
Called for its side effect of writing the sample annotation file;
  returns invisibly NULL.
}
\description{
Write sample annotation into a tab-delimited file to start the Biokit
pipeline
}
\note{
Starting from version 1.0-36, the function checks the input
\code{data.frame} or \code{tbl_df} before writing to the file
}
\examples{

testDf <- data.frame(ID=LETTERS[1:4], 
   GROUP=gl(2,2), 
   FASTQ1=sprintf("\%s_R1.fastq.gz", LETTERS[1:4]),
   FASTQ2=sprintf("\%s_R1.fastq.gz", LETTERS[1:4]))
tmp <- tempfile()
writeBiokitSampleAnnotation(testDf, con=tmp)
readLines(tmp)

}
