% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportEdgeResult.R
\name{writeDgeTablesWithPseudoT}
\alias{writeDgeTablesWithPseudoT}
\title{Write dgeTables with pseudo T statistics}
\usage{
writeDgeTablesWithPseudoT(edgeResult, outdir = getwd())
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{outdir}{Output directory}
}
\value{
\code{NULL}, side effects are used
}
\description{
Write dgeTables with pseudo T statistics
}
\seealso{
\code{\link{dgeTableWithPseudoT}}
}
