% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportEdgeResult.R
\name{writeTruncatedDgeTables}
\alias{writeTruncatedDgeTables}
\title{Write truncated DGE tables}
\usage{
writeTruncatedDgeTables(edgeResult, outdir = getwd())
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{outdir}{Output directory}
}
\value{
\code{NULL}, side effects are used
}
\description{
Write truncated DGE tables
}
