% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{zscoreDGE}
\alias{zscoreDGE}
\title{Calculate mid-p quantile residuals}
\usage{
zscoreDGE(y, design = NULL, contrast = ncol(design))
}
\arguments{
\item{y}{An DGEList object}

\item{design}{Design matrix}

\item{contrast}{Contrast vector

The function is a carbon copy of edgeR:::.zscoreDGE, which is unfortunately not exported
TODO: automatically check whether the function has changed in edgeR in the future}
}
\description{
Calculate mid-p quantile residuals
}
