% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMethods.R
\name{annotate,EdgeObject,character,logical-method}
\alias{annotate,EdgeObject,character,logical-method}
\title{Annotate an EdgeObject}
\usage{
\S4method{annotate}{EdgeObject,character,logical}(object, target, check.target)
}
\arguments{
\item{object}{An EdgeObject.}

\item{target}{Character, target of annotation.}

\item{check.target}{Logical, check whether the target is valid or not.}
}
\description{
Annotate an EdgeObject
}
