% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateMethods.R
\name{annotate,EdgeObject,missing,missing-method}
\alias{annotate,EdgeObject,missing,missing-method}
\title{Annotate an EdgeObject automatically without checking the target}
\usage{
\S4method{annotate}{EdgeObject,missing,missing}(object)
}
\arguments{
\item{object}{An EdgeObject}
}
\description{
Annotate an EdgeObject automatically without checking the target
}
