% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{camera.EdgeResult}
\alias{camera.EdgeResult}
\title{Run CAMERA method using EdgeResult}
\usage{
\method{camera}{EdgeResult}(y, gmtList)
}
\arguments{
\item{y}{A EdgeResult object}

\item{gmtList}{Gene set collections, for example read by
\code{\link[BioQC]{readGmt}}

Note that the EdgeResult object must have a column 'GeneSymbol' in its
\code{fData}.}
}
\description{
Run CAMERA method using EdgeResult
}
