% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{commonDisp-set}
\alias{commonDisp-set}
\alias{commonDisp<-,DGEList,numeric-method}
\alias{commonDisp<-,EdgeObject,numeric-method}
\title{Set common dispersion}
\usage{
commonDisp(object) <- value

\S4method{commonDisp}{DGEList,numeric}(object) <- value

\S4method{commonDisp}{EdgeObject,numeric}(object) <- value
}
\description{
Set common dispersion
}
\section{Functions}{
\itemize{
\item \code{commonDisp<-,DGEList,numeric-method}: Method for DGEList

\item \code{commonDisp<-,EdgeObject,numeric-method}: Method for EdgeObject
}}

