% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{commonDisp<-}
\alias{commonDisp<-}
\alias{commonDisp<-,DGEList,numeric-method}
\alias{commonDisp<-,EdgeObject,numeric-method}
\title{commonDisp-set}
\usage{
commonDisp(object) <- value

\S4method{commonDisp}{DGEList,numeric}(object) <- value

\S4method{commonDisp}{EdgeObject,numeric}(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Numeric value}
}
\description{
Set common dispersion
}
\section{Functions}{
\itemize{
\item \code{commonDisp(object = DGEList) <- value}: Method for DGEList

\item \code{commonDisp(object = EdgeObject) <- value}: Method for EdgeObject

}}
