% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{contrastMatrix-set}
\alias{contrastMatrix-set}
\alias{contrastMatrix<-,EdgeObject,matrix-method}
\title{Assign contrast matrix}
\usage{
contrastMatrix(object) <- value

\S4method{contrastMatrix}{EdgeObject,matrix}(object) <- value
}
\description{
Assign contrast matrix
}
\section{Functions}{
\itemize{
\item \code{contrastMatrix<-,EdgeObject,matrix-method}: Method for EdgeObject
}}

