% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastNames,EdgeObject-method}
\alias{contrastNames,EdgeObject-method}
\title{Extract contrast names from an EdgeObject object}
\usage{
\S4method{contrastNames}{EdgeObject}(object)
}
\arguments{
\item{object}{An EdgeObject object}
}
\description{
Extract contrast names from an EdgeObject object
}
