% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastNames,EdgeResult-method}
\alias{contrastNames,EdgeResult-method}
\title{Extract contrast names from an EdgeResult object}
\usage{
\S4method{contrastNames}{EdgeResult}(object)
}
\arguments{
\item{object}{An EdgeResult object}
}
\description{
Extract contrast names from an EdgeResult object
}
