% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastSampleIndices,EdgeResult,character-method}
\alias{contrastSampleIndices,EdgeResult,character-method}
\title{Extract contrast sample indices}
\usage{
\S4method{contrastSampleIndices}{EdgeResult,character}(object, contrast)
}
\arguments{
\item{object}{An EdgeResult object.}

\item{contrast}{Character, indicating the contrast of interest.}
}
\description{
Extract contrast sample indices
}
