% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastSampleIndices,EdgeResult,integer-method}
\alias{contrastSampleIndices,EdgeResult,integer-method}
\title{Extract contrast sample indices}
\usage{
\S4method{contrastSampleIndices}{EdgeResult,integer}(object, contrast)
}
\description{
Extract contrast sample indices
}
