% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{counts,EdgeObject-method}
\alias{counts,EdgeObject-method}
\title{Return counts in EdgeObject}
\usage{
\S4method{counts}{EdgeObject}(object, filter = TRUE)
}
\arguments{
\item{object}{An EdgeObject}

\item{filter}{Logical, whether filtered matrix (by default) or unfiltered matrix should be returned}
}
\description{
Return counts in EdgeObject
}
\seealso{
\code{\link{filterByCPM}}
}
