% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeObject.R
\name{designContrast}
\alias{designContrast}
\title{Retrieve the design/contrast object}
\usage{
designContrast(edgeObject)
}
\arguments{
\item{edgeObject}{An \code{EdgeObject}}
}
\description{
Retrieve the design/contrast object
}
\examples{
designContrast(exampleEdgeObject())
}
