% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosNGS.R
\name{DesignContrast}
\alias{DesignContrast}
\title{DesignContrast is imported from DesignContrast and exported for convenience}
\description{
DesignContrast is imported from DesignContrast and exported for convenience
}
