% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{designMatrix,EdgeResult-method}
\alias{designMatrix,EdgeResult-method}
\title{Extract design matrix from an EdgeResult object}
\usage{
\S4method{designMatrix}{EdgeResult}(object)
}
\arguments{
\item{object}{An EdgeResult object}
}
\description{
Extract design matrix from an EdgeResult object
}
