% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{DGEListList}
\alias{DGEListList}
\title{Construct a DGEListList object}
\usage{
DGEListList(...)
}
\arguments{
\item{...}{A list of DGEListList objects, can be passed as individual objects
or in a list}
}
\description{
Construct a DGEListList object
}
