% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEListToLongTable.R
\name{DGEListToLongTable}
\alias{DGEListToLongTable}
\title{Convert a DGEList object to a long data.frame containing expression, 
feature annotation, and sample annotation}
\usage{
DGEListToLongTable(x, exprsFun = function(dgeList) cpm(dgeList, log = TRUE))
}
\arguments{
\item{x}{A \code{DGEList} object}

\item{exprsFun}{A function to convert counts to expression data. 
Default: logCPM}
}
\description{
Convert a DGEList object to a long data.frame containing expression, 
feature annotation, and sample annotation
}
\note{
Columns with empty names will be discard.
}
\examples{
mat <- matrix(rnbinom(100, mu=5, size=2), ncol=10)
rownames(mat) <- sprintf("gene\%d", 1:nrow(mat))
y <- edgeR::DGEList(counts=mat, group=rep(1:2, each=5))
DGEListToLongTable(y)

}
