% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{dispGroups,EdgeObject-method}
\alias{dispGroups,EdgeObject-method}
\title{Get display labels of sample groups}
\usage{
\S4method{dispGroups}{EdgeObject}(object)
}
\arguments{
\item{An}{\code{EdgeObject} object}
}
\description{
Get display labels of sample groups
}
