% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EdgeObject-class}
\alias{EdgeObject-class}
\title{EdgeObject argumenting DGEList by including designContrast information}
\description{
EdgeObject argumenting DGEList by including designContrast information
}
\section{Slots}{

\describe{
\item{\code{dgeList}}{A \code{DGEList} object}

\item{\code{designContrast}}{A \code{designContrast} object}
}}

