% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeObject.R
\name{EdgeObject}
\alias{EdgeObject}
\title{Construct an EdgeObject object by a count matrix and DesignContrast}
\usage{
EdgeObject(object, designContrast, ...)
}
\arguments{
\item{object}{A matrix containing counts of features}

\item{designContrast}{A \code{DesignContrast} object}

\item{fData}{A \code{data.frame} containing annotation information for each gene}

\item{pData}{A \code{data.frame} containing annotation information for each sample}

\item{remove.zeros}{Logical, whether to remove rows that have 0 total count}
}
\description{
Construct an EdgeObject object by a count matrix and DesignContrast
}
\examples{
exMat <- matrix(rpois(120, 10), nrow=20, ncol=6)
exGroups <- gl(2,3, labels=c("Group1", "Group2"))
exDesign <- model.matrix(~0+exGroups)
exContrast <- matrix(c(-1,1), ncol=1, 
    dimnames=list(c("Group1", "Group2"), c("Group2.vs.Group1")))
exDescon <- DesignContrast(exDesign, exContrast, groups=exGroups)
exFdata <- data.frame(Identifier=sprintf("Gene\%d", 1:nrow(exMat)))
exPdata <- data.frame(Name=sprintf("Sample\%d", 1:ncol(exMat)),
                     Group=exGroups)
exObj <- EdgeObject(exMat, exDescon)
exObj2 <- EdgeObject(exMat, exDescon, fData=exFdata)
exObj3 <- EdgeObject(exMat, exDescon, 
                     fData=exFdata, pData=exPdata)
            
fData(exObj3)

## note that pData are appended after count information
pData(exObj2)
pData(exObj3)
}
