% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EdgeResult-class}
\alias{EdgeResult-class}
\title{Object that contains test results, dgeTable, and EdgeSigFilter}
\description{
Object that contains test results, dgeTable, and EdgeSigFilter
}
\section{Slots}{

\describe{
\item{\code{dgeGLM}}{A DGEGLM class object that contains GLM test results}

\item{\code{dgeTables}}{A list of dgeTable}

\item{\code{sigFilter}}{Significantly regulated gene filter}
}}

