% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EdgeSigFilter-class}
\alias{EdgeSigFilter-class}
\title{Extends BaseSigFilter to filter genes base on logCPM and LR}
\description{
Extends BaseSigFilter to filter genes base on logCPM and LR
}
\section{Slots}{

\describe{
\item{\code{logCPM}}{Numeric, logCPM threshold (larger values are kept)}
}}

