% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{exportEdgeResult}
\alias{exportEdgeResult}
\title{Export dgeTest results}
\usage{
exportEdgeResult(
  edgeResult,
  outRootDir,
  action = c("ask", "overwrite", "append", "no")
)
}
\arguments{
\item{outRootDir}{Character string, output directory}

\item{action}{Character string, what happens if the output directory exists}

\item{dgeTest}{A \code{EdgeResult} object}
}
\description{
Export dgeTest results
}
