% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{fData<-,EdgeObject,data.frame-method}
\alias{fData<-,EdgeObject,data.frame-method}
\title{Set fData}
\usage{
\S4method{fData}{EdgeObject,data.frame}(object) <- value
}
\arguments{
\item{object}{An EdgeObject}

\item{value}{A \code{data.frame}}
}
\description{
Set fData
}
