% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{filterByCPM}
\alias{filterByCPM}
\title{Filter lowly expressed genes by counts per million (CPM)}
\usage{
filterByCPM(obj, ...)
}
\description{
Filter lowly expressed genes by counts per million (CPM)
}
