% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{groupCol}
\alias{groupCol}
\title{Build DGEList object ready for generalized linear model fitting}
\usage{
groupCol(edgeObj, panel = "Set1")
}
\arguments{
\item{x}{a numerical matrix or any object that can be transformed as such}

\item{design}{design matrix (subject to dimensionality check)}

\item{group}{group vector or factor (see \code{\link{DGEList}})}

\item{genes}{data frame for gene or feature annotation (see \code{\link{DGEList}})}

\item{remove.zeros}{logical, whether or not remove zeros (see \code{\link{DGEList}})}

\item{robust}{Whether robust estimation (Zhou \emph{et al.}, NAR 2014) should be used}

\item{fit}{An \code{DGEGLM} object produced by \code{glmFit}}

\item{contrast}{a vector of contrasts}

\item{fit}{An \code{DGEGLM} object produced by \code{glmFit}}

\item{contrasts}{a contrast matrix. If a vector is given, it is converted into a column vector.}

\item{x}{A matrix of read counts, or any objects that can be converted to a matrix}

\item{design}{Design matrix (subject to dimensionality check)}

\item{contrasts}{Contrast matrix (subject to dimensionality check)}

\item{robust}{Logical, whether robust dispersion estimation should be used or not}
}
\value{
A DGEList object with dispersions estimated, ready for \code{glmFit}.

A data frame containing information about differential expression

A list of data frames containing information about differential expression

A list of data frames reporting differntial gene expression
}
\description{
The function combines the step of object initialization, library size calculation, and dispersion estimation
}
\note{
robust=TRUE is only valid for edgeR version >= 3.6.0

The function checks that the resulting table must contain the following columns: logFC, logCPM, LR, PValue, FDR

robust=TRUE is only valid for edgeR version >= 3.6.0
}
\examples{
expression <- matrix(rnbinom(10000,mu=5,size=2),ncol=4)
design <- cbind(baseline=c(1,1,1,1), x=c(1,1,-1,-1))
edgeObj <- edgeBuild(expression, design)
}
\seealso{
\code{glmFit} to fit a generalized linear model

\code{\link{glmFit}} to produce the input object, \code{\link{edgeTest}} to run the test for a contrast matrix.

\code{\link{edgeTestContrast}} to test one contrast vector

\code{\link{edgeBuild}} to build a DGEList object and to estimate dispersions
\code{\link{glmFit}} to fit a DGEList object against the design matrix
\code{\link{edgeTest}} to derive lists of differentially expressed genes

#'
}
\keyword{internal}

