% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{humanGeneSymbols}
\alias{humanGeneSymbols}
\alias{humanGeneSymbols,DGEList-method}
\alias{humanGeneSymbols,EdgeObject-method}
\title{Get human gene symbols for gene-set enrichment analysis}
\usage{
humanGeneSymbols(object)

\S4method{humanGeneSymbols}{DGEList}(object)

\S4method{humanGeneSymbols}{EdgeObject}(object)
}
\description{
Get human gene symbols for gene-set enrichment analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{DGEList}: Method for DGEList

\item \code{EdgeObject}: Method for EdgeObject
}}

