% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{humanGeneSymbols}
\alias{humanGeneSymbols}
\title{Get human gene symbols for gene-set enrichment analysis}
\usage{
humanGeneSymbols(object)
}
\description{
Get human gene symbols for gene-set enrichment analysis
}
