% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/annotateMethods.R
\name{isAnnotated}
\alias{isAnnotated}
\alias{isAnnotated,EdgeObject-method}
\title{Is the object annotated}
\usage{
isAnnotated(object)

\S4method{isAnnotated}{EdgeObject}(object)
}
\description{
Is the object annotated
}
\section{Methods (by class)}{
\itemize{
\item \code{EdgeObject}: Method for EdgeObject
}}

