% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isUnsetAveExpr}
\alias{isUnsetAveExpr}
\title{Whether the aveExpr filter is set}
\usage{
isUnsetAveExpr(limmaSigFilter)
}
\arguments{
\item{limmaSigFilter}{A \code{LimmaSigFilter} object}
}
\description{
Whether the aveExpr filter is set
}
