% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isUnsetLogCPM}
\alias{isUnsetLogCPM}
\title{Whether the logCPM filter is set}
\usage{
isUnsetLogCPM(edgeSigFilter)
}
\arguments{
\item{edgeSigFilter}{A \code{EdgeSigFilter} object}
}
\description{
Whether the logCPM filter is set
}
