% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isUnsetPosLogFC}
\alias{isUnsetPosLogFC}
\alias{isUnsetNegLogFC}
\alias{isUnsetPValue}
\alias{isUnsetFDR}
\title{Tells whether the threshold was not set}
\usage{
isUnsetPosLogFC(sigFilter)

isUnsetNegLogFC(sigFilter)

isUnsetPValue(sigFilter)

isUnsetFDR(sigFilter)
}
\arguments{
\item{sigFilter}{An SigFilter object}
}
\value{
Logical, whether the thresholds are the default values
}
\description{
Tells whether the threshold was not set
}
