% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{isUnsetSigFilter}
\alias{isUnsetSigFilter}
\title{Whether the EdgeSigFilter is the default one}
\usage{
isUnsetSigFilter(object)
}
\arguments{
\item{object}{An EdgeSigFilter object}
}
\value{
Logical, whether it is unset
}
\description{
Whether the EdgeSigFilter is the default one
}
