% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-slurm-lsf.R
\name{lsfEdgeRcommand}
\alias{lsfEdgeRcommand}
\title{Return the LSF command to run the edgeR script}
\usage{
lsfEdgeRcommand(
  dgeList,
  designContrast,
  outdir = "edgeR_output",
  outfilePrefix = "an-unnamed-project-",
  mps = FALSE,
  limmaVoom = FALSE,
  appendGmt = NULL,
  qos = c("long", "preempty", "short"),
  rootPath = "/apps/rocs/pRED/groups/bioinfo/geneexpression",
  debug = FALSE,
  bsubFile = NULL
)
}
\arguments{
\item{dgeList}{An \code{DGEList} object with \code{counts}, \code{genes},
and \code{samples}}

\item{designContrast}{The DesignContrast object to model the data}

\item{outdir}{Output directory of the edgeR script. Default value
"edgeR_output".}

\item{outfilePrefix}{Prefix of the output files. It can include directories,
e.g. \code{"data/outfile-"}. In case of \code{NULL}, temporary files will be
created.}

\item{mps}{Logical, whether molecular-phenotyping analysis is run.}

\item{limmaVoom}{Logical, whether the limma-voom model is run instead of the edgeR model}

\item{appendGmt}{\code{NULL} or character string, path to an additional GMT
file for gene-set analysis. The option is passed to
\code{\link{edgeRcommand}}.}

\item{qos}{Character, specifying Quality of Service of LSF Available values include \code{long}, \code{short}, and \code{preempty}.}

\item{rootPath}{Character string, the directory of geneexpression scripts, under which \code{bin/ngsDge_edgeR.Rscript} is found.}

\item{debug}{Logical, if \code{TRUE}, the source code of Rscript is used instead of
the installed version. The option is passed to \code{edgeRcommand}.}

\item{bsubFile}{\code{NULL} or character string, file name that contains LSF jobs. If \code{NULL}, a file will be generated within the current directory following the pattern of \code{outfilePrefix.bsub}.

This function wraps the function \code{\link{edgeRcommand}} to return the
command needed to start a LSF job.

It uses \code{outdir} to specify slurm output and error files as in the same
directory of \code{outdir}. And the job name is set as the name of the
output directory.}
}
\description{
Return the LSF command to run the edgeR script
}
\examples{

 mat <- matrix(rnbinom(100, mu=5, size=2), ncol=10)
 rownames(mat) <- sprintf("gene\%d", 1:nrow(mat))
 myFac <- gl(2,5, labels=c("Control", "Treatment"))
 y <- edgeR::DGEList(counts=mat, group=myFac)
 myDesign <- model.matrix(~myFac); colnames(myDesign) <- levels(myFac)
 myContrast <- limma::makeContrasts(Treatment, levels=myDesign)
 myDesCon <- DesignContrast(designMatrix=myDesign, contrastMatrix=myContrast)
 lsfEdgeRcommand(y, designContrast=myDesCon,
     outfilePrefix="test", outdir=tempdir())
 ## remove the bsub file
 file.remove("test.bsub")

}
\seealso{
\code{\link{edgeRcommand}}
}
