% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{mergeAmpliseqRuns}
\alias{mergeAmpliseqRuns}
\title{Merge AmpliSeq runs}
\usage{
mergeAmpliseqRuns(readCountList, barcodeSummaryList, runNames = NULL)
}
\arguments{
\item{readCountList}{A list of read count data.frames}

\item{barcodeSummaryList}{A list of barcode summary data.frames}

\item{runNames}{Character strings, run names. If \code{NULL}, a sequential
number will be given.

The function is used internally to merge several runs into one ExpressionSet
object.}
}
\description{
Merge AmpliSeq runs
}
