% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{modLogCPM}
\alias{modLogCPM}
\alias{modLogCPM,DGEList-method}
\alias{modLogCPM,EdgeObject-method}
\title{Modulated logCPM}
\usage{
modLogCPM(object, ...)

\S4method{modLogCPM}{DGEList}(object, prior.count = 2)

\S4method{modLogCPM}{EdgeObject}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Modulated logCPM
}
\section{Methods (by class)}{
\itemize{
\item \code{DGEList}: Method for DGEList

\item \code{EdgeObject}: Method for EdgeObject
}}

