% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{nContrast,EdgeResult-method}
\alias{nContrast,EdgeResult-method}
\title{Return the number of contrasts}
\usage{
\S4method{nContrast}{EdgeResult}(object)
}
\arguments{
\item{object}{An \code{EdgeResult} object.}
}
\description{
Return the number of contrasts
}
