% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize,EdgeObject-method}
\alias{normalize,EdgeObject-method}
\title{Normalize an EdgeObject}
\usage{
\S4method{normalize}{EdgeObject}(object, method = "RLE", ...)
}
\arguments{
\item{method}{Method passed to \code{\link[edgeR]{calcNormFactors}}.}

\item{obj}{An \code{EdgeObject} object}
}
\description{
Normalize an EdgeObject
}
\section{Methods (by class)}{
\itemize{
\item \code{EdgeObject}: Normalize an EdgeObject by calculating normalization
factors using the given method
}}

