% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{normBoxplot}
\alias{normBoxplot}
\title{Plot distribution of normalized counts}
\usage{
normBoxplot(before.norm, after.norm, ...)
}
\arguments{
\item{edgeObj}{An \code{EdgeObject}}
}
\description{
Plot distribution of normalized counts
}
