% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{normFactors}
\alias{normFactors}
\alias{normFactors,DGEList-method}
\alias{normFactors,EdgeObject-method}
\title{Extract normalisation factors from the object}
\usage{
normFactors(object)

\S4method{normFactors}{DGEList}(object)

\S4method{normFactors}{EdgeObject}(object)
}
\description{
Extract normalisation factors from the object
}
\section{Methods (by class)}{
\itemize{
\item \code{DGEList}: Method for DGEList

\item \code{EdgeObject}: Method for EdgeObject
}}

