% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{pairs.EdgeResult}
\alias{pairs.EdgeResult}
\title{Pairs plot for EdgeResult}
\usage{
\method{pairs}{EdgeResult}(
  x,
  lower.panel = panel.lmSmooth,
  upper.panel = panel.cor,
  freeRelation = TRUE,
  pch = 19,
  ...
)
}
\arguments{
\item{x}{An EdgeResult object.}

\item{lower.panel}{Lower panel, passed to \code{pairs}.}

\item{upper.panel}{Upper panel, passed to \code{pairs}.}

\item{freeRelation}{Logical, whether x- and y-axis shoule have the same range}

\item{pch}{Point symbol}

\item{...}{passed to \code{pairs}

Plot pairwise logFCs}
}
\description{
Pairs plot for EdgeResult
}
\seealso{
\code{\link[graphics]{pairs}}.
}
