% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMethods.R
\name{plotBCV}
\alias{plotBCV}
\alias{plotBCV,DGEList-method}
\alias{plotBCV,EdgeObject-method}
\alias{plotBCV,EdgeResult-method}
\title{Plot BCV}
\usage{
plotBCV(x, ...)

\S4method{plotBCV}{DGEList}(x, ...)

\S4method{plotBCV}{EdgeObject}(x, ...)

\S4method{plotBCV}{EdgeResult}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Other paramters}
}
\description{
Plot BCV
}
\section{Methods (by class)}{
\itemize{
\item \code{plotBCV(DGEList)}: Method for DGEList

\item \code{plotBCV(EdgeObject)}: Method for EdgeObject

\item \code{plotBCV(EdgeResult)}: Method for EdgeResult

}}
