% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMethods.R
\name{plotMDS.EdgeObject}
\alias{plotMDS.EdgeObject}
\title{plotMDS for EdgeObject}
\usage{
\method{plotMDS}{EdgeObject}(x, ...)
}
\arguments{
\item{x}{An EdgeObject object}
}
\description{
plotMDS for EdgeObject
}
