% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{posLogFC}
\alias{posLogFC}
\alias{negLogFC}
\alias{logCPM}
\alias{LR}
\alias{FDR}
\title{Get settings in the significance filter}
\usage{
posLogFC(edgeSigFilter)

negLogFC(edgeSigFilter)

logCPM(edgeSigFilter)

LR(edgeSigFilter)

FDR(edgeSigFilter)
}
\arguments{
\item{edgeSigFilter}{An EdgeSigFilter object}
}
\value{
Numeric values of the thresholds
}
\description{
Get settings in the significance filter
}
