% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{posLogFC}
\alias{posLogFC}
\alias{negLogFC}
\alias{pValue}
\alias{FDR}
\title{Get settings in the significance filter}
\usage{
posLogFC(sigFilter)

negLogFC(sigFilter)

pValue(sigFilter)

FDR(sigFilter)
}
\arguments{
\item{sigFilter}{An SigFilter object}
}
\value{
Numeric values of the thresholds
}
\description{
Get settings in the significance filter
}
