% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prcomp.DGEList}
\alias{prcomp.DGEList}
\title{Principal component analysis of DGEList}
\usage{
\method{prcomp}{DGEList}(x, ntop = NULL, fun = function(x) cpm(x, log =
  TRUE))
}
\arguments{
\item{x}{A \code{DGEList} object}

\item{ntop}{Integer, how many top-variable genes should be used?}

\item{fun}{Function, how to transform counts in the DGEList into data appropriate for PCA? log-cpm is used by default.

If many genes have zero count in all samples, the PCA plot of samples can be sometimes delusive. Therefore, the function
removes such all-zero-count features prior to PCA analysis.}
}
\description{
Principal component analysis of DGEList
}
\examples{
myCounts <- matrix(rnbinom(100, 3, 0.25), nrow=10)
myDgeList <- DGEList(counts=myCounts,
  samples=data.frame(group=gl(5,2)))
myPrcomp <- prcomp(myDgeList)

#' features with zero count in all samples do not contribute to the PCA analysis
myDgeList2 <- DGEList(counts=rbind(myCounts, rep(0, 10)),
  samples=data.frame(group=gl(5,2)))
myPrcomp2 <- prcomp(myDgeList2)
stopifnot(identical(myPrcomp, myPrcomp2))
}
