% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prcomp.DGEList}
\alias{prcomp.DGEList}
\title{Principal component analysis of DGEList}
\usage{
\method{prcomp}{DGEList}(x, ntop = NULL, scale = FALSE,
  fun = function(x) cpmVsn2(x))
}
\arguments{
\item{x}{A \code{DGEList} object}

\item{ntop}{Integer, how many top-variable genes should be used?}

\item{scale}{Logical, whether variance of features should be scaled to 1. Default \code{FALSE}

If many genes have zero count in all samples, the PCA plot of samples can be sometimes delusive. Therefore, the function
removes such all-zero-count features prior to PCA analysis.}

\item{fun}{Function, how to transform counts in the DGEList into data appropriate for PCA? vsn2 transformation of log2-cpm is used by default.}
}
\description{
Principal component analysis of DGEList
}
\examples{
myCounts <- matrix(rnbinom(1000, 3, 0.25), nrow=100)
myDgeList <- DGEList(counts=myCounts,
  samples=data.frame(group=gl(5,2)))
myPrcomp <- prcomp(myDgeList)

#' features with zero count in all samples do not contribute to the PCA analysis
myDgeList2 <- DGEList(counts=rbind(myCounts, rep(0, 10)),
  samples=data.frame(group=gl(5,2)))
myPrcomp2 <- prcomp(myDgeList2)
stopifnot(identical(myPrcomp, myPrcomp2))
}
\seealso{
\code{\link{prcompExprs}}
}
