% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prcompExprs}
\alias{prcompExprs}
\title{Principal component analysis of expression matrix}
\usage{
prcompExprs(matrix, ntop = NULL)
}
\arguments{
\item{matrix}{Numeric matrix. Features in rows and samples in columns.}

\item{ntop}{Integer or NULL. If not \code{NULL}, only \code{ntop} genes with the highest variance are used for the calculation.}
}
\description{
Principal component analysis of expression matrix
}
\examples{
myTestExprs <- matrix(rnorm(1000), ncol=10, byrow=FALSE)
myTestExprs[1:100, 6:10] <- myTestExprs[1:100, 6:10] + 2
myTopPca <- prcompExprs(myTestExprs, ntop=100)
}
