% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoT.R
\name{pseudoTfromPvalue}
\alias{pseudoTfromPvalue}
\title{Convert p-values to t-statistics}
\usage{
pseudoTfromPvalue(p, df, sign, replaceZero = TRUE)
}
\arguments{
\item{p}{Numeric, a numeric vector between 0 and 1.}

\item{df}{Numeric, degree of freedom.}

\item{sign}{Logical or integer, positive numbers or \code{TRUE} are interpreted as positive, and negative numbers or \code{TRUE} are interpreted as negative.}

\item{replaceZero}{Logical, whether small p values or 0 should be replaced by a sufficient small number. Default and recommended: \code{TRUE}}
}
\description{
Convert p-values to t-statistics
}
\examples{
pVals <- 10^(seq(-11,0))
signs <- rep(c(TRUE, FALSE), 6)
tVals <- pseudoTfromPvalue(pVals, 5, sign=signs)
logFCs <- rep(c(1.2,-1.2),6)
tValsLogFCs <- pseudoTfromPvalue(pVals, 5, sign=logFCs)
}
