% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{readAmpliSeq}
\alias{readAmpliSeq}
\title{Read AmpliSeq results into an ExpressionSet object}
\usage{
readAmpliSeq(
  readCountFiles,
  barcodeSummaryFiles,
  runNames = names(readCountFiles)
)
}
\arguments{
\item{readCountFiles}{Character string vector, names of read count files}

\item{barcodeSummaryFiles}{Character string vector, names of barcode summary
files}

\item{runNames}{Character string vector, run names

This function parses read count files as well as barcode summary files, and
organise the data into an \code{\linkS4class{ExpressionSet}} object.}
}
\description{
Read AmpliSeq results into an ExpressionSet object
}
\examples{

countFiles <- system.file("extdata/AmpliSeq_files/",
  sprintf("ReadCountFile-\%d.xls", 1:3), package="ribiosNGS")
barcodeFiles <- system.file("extdata/AmpliSeq_files/",
  sprintf("BarcodeSummaryFile-\%d.xls", 1:3), package="ribiosNGS")
exampleAmpliSeq <- readAmpliSeq(countFiles, barcodeFiles, 
  runNames=c("R1", "R2", "R3"))

}
