% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitAsDGEList.R
\name{readBiokitAsDGEList}
\alias{readBiokitAsDGEList}
\title{Read a Biokit output directory into a DGEList object for downstream analysis}
\usage{
readBiokitAsDGEList(
  dir,
  anno = c("refseq", "ensembl", "gencode"),
  useCollapsedData = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dir}{Biokit output directory}

\item{anno}{Annotation type, either \code{refseq} or \code{ensembl} is
supported}

\item{useCollapsedData}{Logical, \code{FALSE} as default. If set to 
\code{TRUE}, counts are collapsed by gene symbols. This is not recommended
because gene symbols are not stable identifiers.}

\item{verbose}{Logical
The function depends on \code{gct} (\code{gct-ens}) and \code{annot} 
directories of biokit output directory.}
}
\value{
A \code{DGEList} object with count and TPM matrices, sample
  annotation, feature annotation, and an additional \code{BiokitAnno}
  element indicating the annotation type used.
}
\description{
Read a Biokit output directory into a DGEList object for downstream analysis
}
\examples{

##... (TODO: add a mock output directory in testdata)

}
