% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitExpression.R
\name{readBiokitAsDGEList}
\alias{readBiokitAsDGEList}
\title{Read a Biokit output directory into a DGEList object for downstream analysis}
\usage{
readBiokitAsDGEList(
  dir,
  anno = c("refseq", "ensembl"),
  countType = c("count", "uniqCount", "star-count"),
  tpmType = c("tpm", "star-tpm")
)
}
\arguments{
\item{dir}{Biokit output directory}

\item{anno}{Annotation type, either \code{refseq} or \code{ensembl} is
supported}

\item{countType}{Character string, type of count data.}

\item{tpmType}{Character string, type of TPM data.}
}
\description{
Read a Biokit output directory into a DGEList object for downstream analysis
}
\examples{

##... (TODO: add a mock output directory in testdata)

}
