% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitExpression.R
\name{readBiokitAsDGEList}
\alias{readBiokitAsDGEList}
\title{Read a Biokit output directory into a DGEList object for downstream analysis}
\usage{
readBiokitAsDGEList(
  dir,
  anno = c("refseq", "ensembl"),
  useCollapsedData = FALSE
)
}
\arguments{
\item{dir}{Biokit output directory}

\item{anno}{Annotation type, either \code{refseq} or \code{ensembl} is
supported}

\item{useCollapsedData}{Logical, \code{FALSE} as default. If set to 
\code{TRUE}, counts are collapsed by gene symbols. This is not recommended
because gene symbols are not stable identifiers.}
}
\description{
Read a Biokit output directory into a DGEList object for downstream analysis
}
\examples{

##... (TODO: add a mock output directory in testdata)

}
