% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitExpression.R
\name{readBiokitAsDGEList}
\alias{readBiokitAsDGEList}
\title{Read a Biokit output directory into a DGEList object for downstream analysis}
\usage{
readBiokitAsDGEList(dir, anno = c("refseq", "ensembl"),
  countType = c("counts", "uniqCounts", "star-counts"),
  tpmType = c("tpms", "star-tpms"))
}
\arguments{
\item{dir}{Biokit output directory}

\item{anno}{Annotation type, either \code{refseq} or \code{ensembl} is
supported}

\item{type}{count type, either \code{count} or \code{uniqCount} are
supported.}
}
\description{
Read a Biokit output directory into a DGEList object for downstream analysis
}
\examples{

##... (TODO: add a mock output directory in testdata)

}
