% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitExpression.R
\name{readBiokitFeatureAnnotation}
\alias{readBiokitFeatureAnnotation}
\title{Read feature annotation from Biokit directory}
\usage{
readBiokitFeatureAnnotation(dir, anno = c("refseq", "ensembl"))
}
\arguments{
\item{dir}{Character string, a Biokit output directory.}

\item{anno}{Character, indicating the annotation type.}
}
\value{
A \code{data.frame} containing feature annotation, with feature IDs 
  as characters in rownames. The data frame contains following columns 
  depending on the \code{anno} parameter: 
  \enumerate{
    \item FeatureID
    \item GeneID (refseq only) or EnsemblID (ensembl only)
    \item GeneSymbol
    \item mean: mean length
    \item median: median length
    \item longest_isoform: longest isoform
    \item merged: total length of merged exons
  }
  
The function depends on the \code{refseq.annot.gz} (\code{ensembl.annot.gz}) 
and \code{refseq.geneLength.gz} (\code{ensembl.geneLength.gz}) files in the 
biokit directory.
}
\description{
Read feature annotation from Biokit directory
}
\examples{
## TODO add small example files
}
