% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitExpression.R
\name{readBiokitGctFile}
\alias{readBiokitGctFile}
\title{Read GCT files from Biokit output directory}
\usage{
readBiokitGctFile(dir, anno = c("refseq", "ensembl"), type = c("count",
  "rpkm", "uniqCount", "uniqRpkm"))
}
\arguments{
\item{dir}{Biokit output directory}

\item{anno}{Annotation type, either \code{refseq} or \code{ensembl} is supported}

\item{type}{GCT file type, \code{count}, \code{rpkm}, \code{uniqCount} and \code{uniqRpkm} are supported.}
}
\value{
A numeric matrix with the attribute \code{desc} encoding the values in the description column of the GCT format
}
\description{
Read GCT files from Biokit output directory
}
\examples{
##... (TODO: add a mock output directory in testdata)
}
