% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBiokitExpression.R
\name{readBiokitGctFile}
\alias{readBiokitGctFile}
\title{Read GCT files from Biokit output directory}
\usage{
readBiokitGctFile(
  dir,
  anno = c("refseq", "ensembl"),
  type = c("count", "tpm", "count_collapsed", "tpm_collapsed", "log2tpm")
)
}
\arguments{
\item{dir}{Biokit output directory}

\item{anno}{Annotation type, either \code{refseq} or \code{ensembl} is
supported}

\item{type}{GCT file type, \code{count}, \code{tpm}, \code{count_collapsed}, 
\code{tpm_collapsed}, and \code{log2tpm} are supported.}
}
\value{
A numeric matrix with the attribute \code{desc} encoding the values
  in the description column of the GCT format.

The function depends on \code{gct} (in case \code{anno="refseq"}) or
 \code{gct-ens} (in case \code{anno="ensembl"}) sub-directory in the biokit
 output directory.
}
\description{
Read GCT files from Biokit output directory
}
\examples{

##... (TODO: add a mock output directory in testdata)

}
