% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFeatureAnnotationForEdgeR.R
\name{readFeatureAnnotationForEdgeR}
\alias{readFeatureAnnotationForEdgeR}
\title{Read feature annotation for EdgeR pipeline}
\usage{
readFeatureAnnotationForEdgeR(featureNames, file = NULL)
}
\arguments{
\item{featureNames}{Character string, feature names}

\item{file}{A tab-delimited file with header that provides feature annotation}
}
\value{
A \code{data.frame}, with the first column named \code{FeatureName},
  which are the input feature names. The rest columns contain annotations.
}
\description{
Read feature annotation for EdgeR pipeline
}
