% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps-io.R
\name{readMpsnakeAsDGEList}
\alias{readMpsnakeAsDGEList}
\title{Read mpsnake output directory into a DGEList object}
\usage{
readMpsnakeAsDGEList(dir)
}
\arguments{
\item{dir}{Character string, path of mpsnake pipeline directory (or the \code{results} subdirectory).}
}
\value{
A \code{DGEList} object containing counts, gene, and sample annotation
}
\description{
Read mpsnake output directory into a DGEList object
}
\examples{
mpsnakeDir <- system.file("extdata/mpsnake-minimal-outdir", "ribiosNGS")
mpsDgeList <- readMpsnakeAsDGEList(mpsnakeDir)

## equivalent
mpsnakeResDir <- file.path(mpsnakeDir, "results")
mpsDgeList <- readMpsnakeAsDGEList(mpsnakeResDir)
}
