% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosNGS.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DGEList}
\alias{DesignContrast}
\alias{cpm}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{edgeR}{\code{\link[edgeR]{cpm}}, \code{\link[edgeR]{DGEList}}}

  \item{ribiosExpression}{\code{\link[ribiosExpression]{DesignContrast}}}
}}

