% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosNGS.R
\docType{package}
\name{ribiosNGS-package}
\alias{ribiosNGS}
\alias{ribiosNGS-package}
\title{ribiosNGS: Utilities for next-generation sequencing data analysis in ribios}
\description{
NGS tasks in ribios, primarily differetial gene expression
    analysis.
}
\details{
ribiosNGS package provides functionalities used for next-generation sequencing data
}
\keyword{internal}
