% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosNGS-package.R
\docType{package}
\name{ribiosNGS-package}
\alias{ribiosNGS-package}
\title{ribiosNGS package provides functionalities used for next-generation sequencing data}
\description{
Provides data structures and functions for next-generation sequencing gene expression analysis
}
\keyword{package}
