% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{sampleNames,DGEList-method}
\alias{sampleNames,DGEList-method}
\title{Return sample names from a DGEList object}
\usage{
\S4method{sampleNames}{DGEList}(object)
}
\arguments{
\item{object}{A DGEList object}
}
\description{
Return sample names from a DGEList object
}
