% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{setCommonDispIfMissing}
\alias{setCommonDispIfMissing}
\alias{setCommonDispIfMissing,DGEList,numeric-method}
\alias{setCommonDispIfMissing,EdgeObject,numeric-method}
\title{Set common dispersion if missing}
\usage{
setCommonDispIfMissing(object, value)

\S4method{setCommonDispIfMissing}{DGEList,numeric}(object, value)

\S4method{setCommonDispIfMissing}{EdgeObject,numeric}(object, value)
}
\description{
Set common dispersion if missing
}
\section{Methods (by class)}{
\itemize{
\item \code{object = DGEList,value = numeric}: Method for DGEList

\item \code{object = EdgeObject,value = numeric}: Method for EdgeObject
}}

