% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,SigFilter-method}
\alias{show,SigFilter-method}
\title{Show an SigFilter object}
\usage{
\S4method{show}{SigFilter}(object)
}
\arguments{
\item{object}{An SigFilter object}
}
\description{
Show an SigFilter object
}
