% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigFilter<-}
\alias{sigFilter<-}
\title{Replace the EdgeSigFilter of an EdgeResult}
\usage{
sigFilter(edgeResult) <- value
}
\arguments{
\item{edgeResult}{An EdgeResult object}

\item{value}{An EdgeSigFilter object}
}
\value{
An updated \code{EdgeResult} object
}
\description{
Replace the EdgeSigFilter of an EdgeResult
}
