% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigFilter<-}
\alias{sigFilter<-}
\title{Replace the SigFilter of an CountDgeResult}
\usage{
sigFilter(countDgeResult) <- value
}
\arguments{
\item{countDgeResult}{An \code{EdgeResult} or \code{LimmaVoomResult} object}

\item{value}{An SigFilter object}
}
\value{
An updated \code{countDgeResult} object
}
\description{
Replace the SigFilter of an CountDgeResult
}
