% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigFilter}
\alias{sigFilter}
\title{Return the SigFilter in use}
\usage{
sigFilter(edgeResult)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}
}
\value{
An \code{SigFilter} object
}
\description{
Return the SigFilter in use
}
