% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{SigFilter}
\alias{SigFilter}
\alias{EdgeSigFilter}
\alias{LimmaSigFilter}
\alias{sigFilter}
\title{Build a SigFilter}
\usage{
SigFilter(logFC, posLogFC, negLogFC, pValue, FDR)
}
\arguments{
\item{logFC}{Missing or positive numeric}

\item{posLogFC}{Missing or positive numeric}

\item{negLogFC}{Missing or negative numeric}

\item{pValue}{Missing or numeric between 0 and 1}

\item{FDR}{Missing or numeric between 0 and 1}
}
\value{
A \code{SigFilter} object
}
\description{
Build a SigFilter
}
\examples{
SigFilter()
SigFilter(logFC=2)
SigFilter(negLogFC=-1)
SigFilter(FDR=0.05)
}
