% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{sigGene}
\alias{sigGene}
\alias{sigPosGene}
\alias{sigNegGene}
\title{Return significantly regulated genes}
\usage{
sigGene(edgeResult, contrast, value = "GeneID")

sigPosGene(edgeResult, contrast, value = "GeneID")

sigNegGene(edgeResult, contrast, value = "GeneID")
}
\arguments{
\item{edgeResult}{An EdgeResult object}

\item{contrast}{Character, contrast(s) of interest}

\item{value}{Character, type of identifier returned}
}
\value{
A vector of identifiers
}
\description{
Return significantly regulated genes
}
\section{Functions}{
\itemize{
\item \code{sigPosGene}: Only return positively significantly regulated genes

\item \code{sigNegGene}: Only return negatively significantly regulated genes
}}

\note{
TODO fix: add InputFeature
}
