% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-slurm-lsf.R
\name{slurmEdgeRcommand}
\alias{slurmEdgeRcommand}
\title{Return the SLURM command to run the edgeR script}
\usage{
slurmEdgeRcommand(
  dgeList,
  designContrast,
  outdir = "edgeR_output",
  outfilePrefix = "an-unnamed-project-",
  mps = FALSE,
  limmaVoom = FALSE,
  appendGmt = NULL,
  qos = c("3h", "1d", "3d", "15d", "interactive", "preempt"),
  params = "",
  rootPath = "/apps/rocs/pRED/groups/bioinfo/geneexpression",
  debug = FALSE
)
}
\arguments{
\item{dgeList}{An \code{DGEList} object with \code{counts}, \code{genes},
and \code{samples}}

\item{designContrast}{The DesignContrast object to model the data}

\item{outdir}{Output directory of the edgeR script. Default value
"edgeR_output".}

\item{outfilePrefix}{Prefix of the output files. It can include directories,
e.g. \code{"data/outfile-"}. In case of \code{NULL}, temporary files will be
created.}

\item{mps}{Logical, whether molecular-phenotyping analysis is run.}

\item{limmaVoom}{Logical, whether the limma-voom model is run instead of the edgeR model}

\item{appendGmt}{\code{NULL} or character string, path to an additional GMT
file for gene-set analysis. The option is passed to
\code{\link{edgeRcommand}}.}

\item{qos}{Character, specifying Quality of Service of Slurm. Available values include \code{short} (recommended default, running time cannot exceed 3 hours), \code{interactive} (useful if you wish to get the results from an interactive session), and \code{normal} (useful if the job is expected to run more than three hours.)
using \code{srun} and the 'interaction' queue of jobs instead of using
\code{sbatch}.}

\item{params}{Character, further parameters to pass to sbatch, for instance "--partition ANOTHER_PARITION"}

\item{rootPath}{Character string, the directory of geneexpression scripts, under which \code{bin/ngsDge_edgeR.Rscript} is found.}

\item{debug}{Logical, if \code{TRUE}, the source code of Rscript is used instead of
  the installed version. The option is passed to \code{edgeRcommand}.

This function wraps the function \code{\link{edgeRcommand}} to return the
command needed to start a SLURM job.

It uses \code{outdir} to specify slurm output and error files as in the same
directory of \code{outdir}. And the job name is set as the name of the
output directory.}
}
\value{
A character string containing the SLURM \code{sbatch} command to
  submit the edgeR analysis job.
}
\description{
Return the SLURM command to run the edgeR script
}
\examples{

 mat <- matrix(rnbinom(100, mu=5, size=2), ncol=10)
 rownames(mat) <- sprintf("gene\%d", 1:nrow(mat))
 myFac <- gl(2,5, labels=c("Control", "Treatment"))
 y <- edgeR::DGEList(counts=mat, group=myFac)
 myDesign <- model.matrix(~myFac); colnames(myDesign) <- levels(myFac)
 myContrast <- limma::makeContrasts(Treatment, levels=myDesign)
 myDesCon <- DesignContrast(designMatrix=myDesign,
                            contrastMatrix=myContrast)
 mytempdir <- tempdir()
 slurmEdgeRcommand(y, myDesCon, outfilePrefix="test", outdir=mytempdir)
}
\seealso{
\code{\link{edgeRcommand}}
}
