% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.DGEList.R
\name{split.DGEList}
\alias{split.DGEList}
\alias{splitDGEList}
\title{Split a DGEList object by a factor of samples (default) or genes}
\usage{
\method{split}{DGEList}(x, f, drop = FALSE, bySample = TRUE, sampleDropLevels = TRUE, ...)

splitDGEList(x, f, drop = FALSE, bySample = TRUE, sampleDropLevels = TRUE, ...)
}
\arguments{
\item{x}{A \code{DGEList} object}

\item{f}{A factor vector. Other types will be coereced into factors.}

\item{drop}{Not used now}

\item{bySample}{Logical, if \code{TRUE}, the samples are split. Otherwise, 
genes are split.}

\item{sampleDropLevels}{Logical, if \code{TRUE}, unused levels in factors 
in the sample annotation are dropped}

\item{...}{Not used so far.}
}
\description{
Split a DGEList object by a factor of samples (default) or genes
}
\section{Functions}{
\itemize{
\item \code{splitDGEList()}: A wrapper of split.DGEList

}}
\examples{
y1 <- matrix(rnbinom(1000, mu=5, size=2), ncol=4)
genes1 <- data.frame(GeneSymbol=sprintf("Gene\%d", 1:nrow(y1)),
  GeneType=gl(5,50))
rownames(y1) <- rownames(genes1) <- 1:nrow(y1)
anno1 <- data.frame(treatment=gl(2,2, labels=c("ctrl", "tmt")),
    donor=factor(rep(c(1,2), each=2)))
d1 <- DGEList(counts=y1, genes=genes1, samples=anno1)

d1SampleSplit <- split(d1, d1$samples$donor)
d1GeneSplit <- split(d1, d1$genes$GeneType, bySample=FALSE)

}
