% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svaseqRemove.R
\name{svaseqRemove}
\alias{svaseqRemove}
\title{Detect surrogate variables from count data and remove their effects VSN-transformed counts}
\usage{
svaseqRemove(dgeList, design, nullModel, verbose = FALSE, offset)
}
\arguments{
\item{dgeList}{An \code{DGEList} object}

\item{design}{Design matrix}

\item{nullModel}{Null model matrix}

\item{verbose}{Logical}

\item{offset}{If provided, it is passed to \code{pcaScores}.

In case no significant surrogate variables are detected, PCA analysis is 
applied to the vsn-transformed matrix.}
}
\value{
A list with following items
\itemize{
  \item{sva}{Results of \code{svaseq}}
  \item{vsnFit}{Fit object of vsn}
  \item{vsnMat}{Fitted matrix of vsn}
  \item{vsnBatchRemoved}{Fitted matrix of vsn, with surrogates' effect 
       removed}
  \item{vsnBatchRemovedPca}{PCA object derived from \code{vsnBatchRemoved}}
  \item{vsnBatchRemovedPcaScores}{PCA scores with annotations}
  \item{designWithSV}{Design matrix with surrogates variables appended
       if any}
}
}
\description{
Detect surrogate variables from count data and remove their effects VSN-transformed counts
}
\note{
This function needs to be harmonized with the other SVA functions.
 The reason is that svaseq was for a long time not stable until recently.
 Therefore this function is written later, and unfortunately the 
 outcome is not harmonized yet.
}
\examples{
y1org <- matrix(rnbinom(4000, mu=5, size=2), ncol=8)
genes1 <- data.frame(GeneSymbol=sprintf("Gene\%d", 1:nrow(y1org)))
y1 <- y1org
y1[30:120, 4:7] <- y1[30:120, 4:7]+9 ## mimicking batch effect
rownames(y1org) <- rownames(y1) <- rownames(genes1) <- 1:nrow(y1)
anno1 <- data.frame(treatment=gl(2,4, labels=c("ctrl", "tmt")),
    donor=factor(rep(c(1,2), 4)))
d1 <- DGEList(counts=y1, genes=genes1, samples=anno1)
d2 <- DGEList(counts=y1org, genes=genes1, samples=anno1)

design <- model.matrix(~treatment+donor, data=d1$samples)
nullModel <-  model.matrix(~donor, data=d1$samples)
d1VsnSvaRes <- svaseqRemove(d1, design, nullModel)
d2VsnSvaRes <- svaseqRemove(d2, design, nullModel)
}
\seealso{
\link[ribiosPlot]{pcaScores}
}
