% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{tagwiseBCV}
\alias{tagwiseBCV}
\alias{tagwiseBCV,DGEList-method}
\alias{tagwiseBCV,EdgeResult-method}
\title{Tagwise biological coefficients of variance}
\usage{
tagwiseBCV(x)

\S4method{tagwiseBCV}{DGEList}(x)

\S4method{tagwiseBCV}{EdgeResult}(x)
}
\arguments{
\item{x}{An object}
}
\description{
Tagwise biological coefficients of variance
}
\section{Methods (by class)}{
\itemize{
\item \code{tagwiseBCV(DGEList)}: method for DGEList

\item \code{tagwiseBCV(EdgeResult)}: method for EdgeResult

}}
