% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{testGLM}
\alias{testGLM}
\alias{testGLM,EdgeObject,DGEGLM-method}
\title{Test GLM}
\usage{
testGLM(object, fit)

\S4method{testGLM}{EdgeObject,DGEGLM}(object, fit)
}
\arguments{
\item{object}{An object}

\item{fit}{A fit object}
}
\description{
Test GLM
}
\section{Methods (by class)}{
\itemize{
\item \code{testGLM(object = EdgeObject, fit = DGEGLM)}: Method for EdgeObject and DGEGLM.

}}
