% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{trendedBCV}
\alias{trendedBCV}
\alias{trendedBCV,DGEList-method}
\alias{trendedBCV,EdgeResult-method}
\title{Trended biological coefficients of variance}
\usage{
trendedBCV(x)

\S4method{trendedBCV}{DGEList}(x)

\S4method{trendedBCV}{EdgeResult}(x)
}
\arguments{
\item{x}{An object}
}
\description{
Trended biological coefficients of variance
}
\section{Methods (by class)}{
\itemize{
\item \code{DGEList}: method for DGEList

\item \code{EdgeResult}: method for EdgeResult
}}

