% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{update.SigFilter}
\alias{update.SigFilter}
\alias{`logFC<-`}
\alias{`negLogFC<-`}
\alias{`aveExpr<-`}
\alias{`pValue<-`}
\alias{`FDR<-`}
\alias{posLogFC<-}
\alias{negLogFC<-}
\alias{logFC<-}
\alias{aveExpr<-}
\alias{pValue<-}
\alias{FDR<-}
\title{Update SigFilter}
\usage{
\method{update}{SigFilter}(object, logFC, posLogFC, negLogFC, aveExpr, pValue, FDR, ...)

posLogFC(object) <- value

negLogFC(object) <- value

logFC(object) <- value

aveExpr(object) <- value

pValue(object) <- value

FDR(object) <- value
}
\arguments{
\item{object}{An \code{SigFilter} object}

\item{logFC}{Numeric, logFC filter value, optional.}

\item{posLogFC}{Numeric, positive logFC filter value, optional.}

\item{negLogFC}{Numeric, negative logFC filter value, optional.}

\item{aveExpr}{Numeric, AveExpr filter value, optional.}

\item{pValue}{Numeric, pValue filter value, optional}

\item{FDR}{Numeric, FDR filter value, optional}

\item{...}{not used now}

\item{value}{Numeric, vssigned threshold value}
}
\value{
An updated \code{SigFilter} object.
}
\description{
Update SigFilter
}
\section{Functions}{
\itemize{
\item \code{posLogFC<-}: Updates the posLogFC threshold value

\item \code{negLogFC<-}: Updates the negLogFC threshold value

\item \code{logFC<-}: Updates the posLogFC threshold value

\item \code{aveExpr<-}: Updates the aveExpr threshold value

\item \code{pValue<-}: Updates the pValue threshold value

\item \code{FDR<-}: Updates the FDR threshold value
}}

