% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR-funcs.R
\name{updateSigFilter}
\alias{updateSigFilter}
\title{Update the SigFilter}
\usage{
updateSigFilter(countDgeResult, logFC, posLogFC, negLogFC, pValue, FDR, ...)
}
\arguments{
\item{countDgeResult}{An \code{EdgeResult} or \code{LimmaVoomResult} object}

\item{logFC}{Numeric}

\item{posLogFC}{Numeric}

\item{negLogFC}{Numeric}

\item{pValue}{Numeric}

\item{FDR}{Numeric}

\item{...}{Other parameters, now used ones including \code{aveExpr} (for LimmaSigFilter) and \code{logCPM} (for EdgeSigFilter)}
}
\value{
An updated \code{CountDgeResult} object with updated \code{SigFilter}
}
\description{
Update the SigFilter
}
