% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMethods.R
\name{volcanoPlot}
\alias{volcanoPlot}
\alias{volcanoPlot,EdgeResult-method}
\title{Volcano plot}
\usage{
volcanoPlot(object, ...)

\S4method{volcanoPlot}{EdgeResult}(
  object,
  contrast = NULL,
  freeRelation = FALSE,
  colramp = ribiosPlot::heat,
  multipage = FALSE,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Volcano plot
}
\section{Methods (by class)}{
\itemize{
\item \code{EdgeResult}: Method for EdgeResult
}}

