% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{voomSVA,DGEList,matrix-method}
\alias{voomSVA,DGEList,matrix-method}
\title{Detect surrogate variables from DGEList}
\usage{
\S4method{voomSVA}{DGEList,matrix}(object, design)
}
\arguments{
\item{object}{A DGEList object}

\item{design}{Design matrix}
}
\value{
A new DGEList object, including new items in the list: \code{voom}, \code{sv}, \code{designMatrix},
\code{designMatrixWithSV}, and \code{voomSVRemoved}.
}
\description{
Detect surrogate variables from DGEList
}
