% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{voomSVA}
\alias{voomSVA}
\alias{voomSVA,matrix,matrix-method}
\alias{voomSVA,DGEList,matrix-method}
\alias{voomSVA,DGEList,formula-method}
\title{Run SVA on a count matrix transformed by voom}
\usage{
voomSVA(object, design, ...)

\S4method{voomSVA}{matrix,matrix}(object, design)

\S4method{voomSVA}{DGEList,matrix}(object, design)

\S4method{voomSVA}{DGEList,formula}(object, design)
}
\arguments{
\item{object}{A count matrix}

\item{design}{Design matrix or formula}

\item{...}{Other parameters}
}
\value{
SV matrix
}
\description{
Run SVA on a count matrix transformed by voom
}
\section{Methods (by class)}{
\itemize{
\item \code{voomSVA(object = matrix, design = matrix)}: Method for count matrix and design matrix

\item \code{voomSVA(object = DGEList, design = matrix)}: Method for DGEList and design matrix

\item \code{voomSVA(object = DGEList, design = formula)}: Method for count matrix and design formula

}}
\examples{
set.seed(1887)
exCounts <- matrix(rpois(12000, 10), nrow=2000, ncol=6)
exCounts[1:100, 2:3] <- exCounts[1:100,2:3]+20
exDesign <- model.matrix(~gl(2,3))
voomSVA(exCounts, design=exDesign)
}
