% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportEdgeResult.R
\name{writeDgeTables}
\alias{writeDgeTables}
\title{Write DGE tables}
\usage{
writeDgeTables(edgeResult, outdir = getwd())
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{outdir}{Output directory}
}
\value{
\code{NULL}, side effects are used
}
\description{
Write DGE tables
}
