% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{zscoreDGE}
\alias{zscoreDGE}
\title{Calculate mid-p quantile residuals}
\usage{
zscoreDGE(y, design = NULL, contrast = ncol(design))
}
\arguments{
\item{y}{An DGEList object}

\item{design}{Design matrix}

\item{contrast}{Contrast vector

The function is a carbon copy of edgeR:::.zscoreDGE, which is unfortunately
not exported}
}
\description{
Calculate mid-p quantile residuals
}
\examples{

dgeMatrix <- matrix(rpois(1200, 10), nrow=200)
dgeList <- DGEList(dgeMatrix)
dgeList <- estimateCommonDisp(dgeList)
dgeDesign <- model.matrix(~gl(2,3))
dgeZscore <- zscoreDGE(dgeList, dgeDesign, contrast=c(0,1))
head(dgeZscore)

}
